🚀 fix: PLT.1-DB.1 マージコンフリクト解消
