+++ scripts/sql-to-seed-prisma-fixed.py
@@ -44,7 +44,7 @@ def parse_sql_file(path: Path) -> Tuple[str, List[str], List[List[Any]]]:
         raise ValueError(f"INSERT文が見つかりません: {path}")
 
     table = match.group(1)
-    columns = [c.strip() for c in match.group(2).split(",")]
+    columns = [c.strip() for c in match.group(2).split(',')]
 
     values_block = match.group(3)
 
@@ -64,17 +64,17 @@ def parse_sql_file(path: Path) -> Tuple[str, List[str], List[List[Any]]]:
         else:
             escape = False
 
-        if ch == "(" and not in_quote:
+        if ch == '(' and not in_quote:
             if depth == 0:
                 buf = []
             else:
                 buf.append(ch)
             depth += 1
             continue
-        if ch == ")" and not in_quote:
+        if ch == ')' and not in_quote:
             depth -= 1
             if depth == 0:
-                row_texts.append("".join(buf).strip())
+                row_texts.append(''.join(buf).strip())
                 continue
             else:
                 buf.append(ch)
@@ -88,11 +88,11 @@ def parse_sql_file(path: Path) -> Tuple[str, List[str], List[List[Any]]]:
         fields: List[Any] = []
         for val in re.split(r",(?=(?:[^']*'[^']*')*[^']*$)", row_text):
             val = val.strip()
-            if val.upper() == "NULL":
+            if val.upper() == 'NULL':
                 fields.append(None)
-            elif val.upper() == "TRUE":
+            elif val.upper() == 'TRUE':
                 fields.append(True)
-            elif val.upper() == "FALSE":
+            elif val.upper() == 'FALSE':
                 fields.append(False)
             else:
                 if val.startswith("'") and val.endswith("'"):
@@ -107,11 +107,11 @@ def parse_sql_file(path: Path) -> Tuple[str, List[str], List[List[Any]]]:
 def format_value(value: Any) -> str:
     """TypeScript用の値表現に変換"""
     if value is None:
-        return "null"
+        return 'null'
     if value is True:
-        return "true"
+        return 'true'
     if value is False:
-        return "false"
+        return 'false'
     if isinstance(value, str) and re.match(r"^-?\d+(\.\d+)?$", value):
         return value
     if isinstance(value, str):
@@ -125,15 +125,15 @@ def generate_ts(table: str, columns: List[str], rows: List[List[Any]]) -> str:
     model_prop = to_prisma_property(table)
     lines.append(f"console.log('📊 {table}データを投入中...')")
     lines.append(f"await prisma.{model_prop}.createMany({{")
-    lines.append("  data: [")
+    lines.append('  data: [')
     for row in rows:
-        lines.append("    {")
+        lines.append('    {')
         for col, val in zip(columns, row):
             lines.append(f"      {col}: {format_value(val)},")
-        lines.append("    },")
-    lines.append("  ],")
-    lines.append("})")
-    return "\n".join(lines)
+        lines.append('    },')
+    lines.append('  ],')
+    lines.append('})')
+    return '\n'.join(lines)
 
 
 def main(sql_dir: str, output_file: str) -> None:
@@ -141,44 +141,43 @@ def main(sql_dir: str, output_file: str) -> None:
     out_path = Path(output_file)
 
     snippets: List[str] = []
-    for sql_file in sorted(sql_path.glob("*_sample_data.sql")):
+    for sql_file in sorted(sql_path.glob('*_sample_data.sql')):
         table, columns, rows = parse_sql_file(sql_file)
         snippet = generate_ts(table, columns, rows)
         snippets.append(f"// {sql_file.name}\n{snippet}\n")
 
     header = [
-        "// Auto generated by sql-to-seed-prisma-fixed.py",
+        '// Auto generated by sql-to-seed-prisma-fixed.py',
         "import { PrismaClient } from '@prisma/client'",
-        "",
-        "const prisma = new PrismaClient()",
-        "",
-        "export async function runSampleSeed() {",
+        '',
+        'const prisma = new PrismaClient()',
+        '',
+        'export async function runSampleSeed() {',
         "  console.log('🌱 データベースの初期データ投入を開始します...')",
     ]
     footer = [
         "  console.log('✅ 初期データ投入が完了しました！')",
-        "}",
-        "",
-        "if (require.main === module) {",
-        "  runSampleSeed().then(() => prisma.$disconnect())",
-        "}",
+        '}',
+        '',
+        'if (require.main === module) {',
+        '  runSampleSeed().then(() => prisma.$disconnect())',
+        '}',
     ]
 
     # Ensure output directory exists
     out_path.parent.mkdir(parents=True, exist_ok=True)
 
-    with out_path.open("w", encoding="utf-8") as f:
-        f.write("\n".join(header) + "\n")
+    with out_path.open('w', encoding='utf-8') as f:
+        f.write('\n'.join(header) + '\n')
         for snippet in snippets:
-            f.write("  " + snippet.replace("\n", "\n  ") + "\n")
-        f.write("\n".join(footer) + "\n")
+            f.write('  ' + snippet.replace('\n', '\n  ') + '\n')
+        f.write('\n'.join(footer) + '\n')
 
 
-if __name__ == "__main__":
+if __name__ == '__main__':
     if len(sys.argv) not in (2, 3):
-        print("Usage: python sql-to-seed-prisma-fixed.py <sql_dir> [output_file]")
+        print('Usage: python sql-to-seed-prisma-fixed.py <sql_dir> [output_file]')
         sys.exit(1)
     sql_dir = sys.argv[1]
-    out_file = sys.argv[2] if len(sys.argv) == 3 else "src/database/prisma/seed.ts"
+    out_file = sys.argv[2] if len(sys.argv) == 3 else 'src/database/prisma/seed.ts'
     main(sql_dir, out_file)
-
