+++ src/database/prisma/seed.ts
@@ -1,3481 +1,549 @@
-// Auto generated by sql-to-seed-prisma-fixed.py
+// 要求仕様ID: PLT.1-DB.1 - データベース初期データ投入
 import { PrismaClient } from '@prisma/client'
 
 const prisma = new PrismaClient()
 
-export async function runSampleSeed() {
+async function main() {
   console.log('🌱 データベースの初期データ投入を開始します...')
-  // HIS_AuditLog_sample_data.sql
-  console.log('📊 HIS_AuditLogデータを投入中...')
-  await prisma.auditLog.createMany({
-    data: [
-      {
-        id: "audit_001",
-        user_id: "emp_001",
-        session_id: "sess_abc123",
-        action_type: "LOGIN",
-        target_table: null,
-        target_id: null,
-        old_values: null,
-        new_values: null,
-        ip_address: "192.168.1.100",
-        user_agent: "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36",
-        result_status: "SUCCESS",
-        error_message: null,
-        execution_time_ms: 150,
-        is_deleted: false,
-        tenant_id: "tenant_001",
-        created_at: null,
-        updated_at: null,
-        created_by: "system",
-        updated_by: "system",
-      },
-      {
-        id: "audit_002",
-        user_id: "emp_001",
-        session_id: "sess_abc123",
-        action_type: "UPDATE",
-        target_table: "MST_Employee",
-        target_id: "emp_001",
-        old_values: "{"name": "田中太郎", "email": "tanaka@example.com"}",
-        new_values: "{"name": "田中太郎", "email": "tanaka.new@example.com"}",
-        ip_address: "192.168.1.100",
-        user_agent: "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36",
-        result_status: "SUCCESS",
-        error_message: null,
-        execution_time_ms: 250,
-        is_deleted: false,
-        tenant_id: "tenant_001",
-        created_at: null,
-        updated_at: null,
-        created_by: "emp_001",
-        updated_by: "emp_001",
-      },
-    ],
+
+  // 1. テナントマスターデータ
+  console.log('🏢 テナントマスターデータを投入中...')
+  const tenant = await prisma.tenant.upsert({
+    where: { tenant_id: 'TENANT001' },
+    update: {},
+    create: {
+      tenant_id: 'TENANT001',
+      tenant_code: 'DEFAULT',
+      tenant_name: 'デフォルトテナント',
+      tenant_short_name: 'DEFAULT',
+      tenant_type: 'ENTERPRISE',
+      domain_name: 'skill-report.local',
+      subdomain: 'default',
+      timezone: 'Asia/Tokyo',
+      locale: 'ja-JP',
+      currency_code: 'JPY',
+      status: 'ACTIVE',
+      max_users: 1000,
+      code: 'TENANT001',
+      name: 'デフォルトテナント',
+    },
   })
-  
-  // HIS_NotificationLog_sample_data.sql
-  console.log('📊 HIS_NotificationLogデータを投入中...')
-  await prisma.notificationLog.createMany({
-    data: [
-      {
-        id: "NL001",
-        tenant_id: "TENANT001",
-        notification_id: "NOTIF001",
-        setting_id: "NS001",
-        template_id: "NT001",
-        notification_type: "EMAIL",
-        recipient_type: "USER",
-        recipient_address: "yamada.taro@company.com",
-        subject: "【スキル更新】山田太郎さんのスキル情報が更新されました",
-        message_body: "山田太郎さん
-  
-  以下のスキル情報が更新されました。
-  
-  スキル名: Java
-  更新日時: 2025-06-01 10:30:00
-  更新者: 佐藤花子
-  
-  詳細は以下のリンクからご確認ください。
-  https://system.company.com/skills/123
-  
-  ※このメールは自動送信されています。",
-        message_format: "PLAIN",
-        send_status: "SUCCESS",
-        send_attempts: 1,
-        max_retry_count: 3,
-        scheduled_at: "2025-06-01 10:30:00",
-        sent_at: "2025-06-01 10:30:15",
-        delivered_at: "2025-06-01 10:30:45",
-        opened_at: "2025-06-01 11:15:30",
-        response_code: 250,
-        response_message: "Message accepted for delivery",
-        error_details: null,
-        integration_config_id: "IC003",
-        priority_level: "MEDIUM",
-        is_deleted: null,
-      },
-      {
-        id: "NL002",
-        tenant_id: "TENANT001",
-        notification_id: "NOTIF002",
-        setting_id: "NS002",
-        template_id: "NT002",
-        notification_type: "SLACK",
-        recipient_type: "CHANNEL",
-        recipient_address: "#notifications",
-        subject: null,
-        message_body: ":warning: *目標期限のお知らせ* :warning:
-  
-  山田太郎さんの目標「Java認定資格取得」の期限が近づいています。
-  
-  • 期限: 2025-06-30
-  • 残り日数: 29日
-  • 進捗率: 75%
-  
-  <https://system.company.com/goals/456|詳細を確認する>",
-        message_format: "MARKDOWN",
-        send_status: "FAILED",
-        send_attempts: 3,
-        max_retry_count: 3,
-        scheduled_at: "2025-06-01 09:00:00",
-        sent_at: null,
-        delivered_at: null,
-        opened_at: null,
-        response_code: 404,
-        response_message: "channel_not_found",
-        error_details: "{"error": "channel_not_found", "details": "The specified channel does not exist or the bot is not a member"}",
-        integration_config_id: "IC001",
-        priority_level: "HIGH",
-        is_deleted: null,
-      },
-    ],
-  })
-  
-  // HIS_TenantBilling_sample_data.sql
-  console.log('📊 HIS_TenantBillingデータを投入中...')
-  await prisma.tenantBilling.createMany({
-    data: [
-      {
-        id: "TB001",
-        tenant_id: "TENANT001",
-        billing_period_start: "2025-05-01",
-        billing_period_end: "2025-05-31",
-        billing_type: "MONTHLY",
-        plan_id: "PLAN_STANDARD",
-        plan_name: "スタンダードプラン",
-        base_amount: 50000.0,
-        usage_amount: 15000.0,
-        additional_amount: 5000.0,
-        discount_amount: 3000.0,
-        subtotal_amount: 67000.0,
-        tax_rate: 0.1,
-        tax_amount: 6700.0,
-        total_amount: 73700.0,
-        currency_code: "JPY",
-        usage_details: "{"users": 25, "storage_gb": 150, "api_calls": 50000}",
-        billing_status: "PAID",
-        invoice_number: "INV-2025-05-001",
-        invoice_date: "2025-06-01",
-        due_date: "2025-06-30",
-        paid_date: "2025-06-15",
-        payment_method: "CREDIT_CARD",
-        notes: "5月分月額利用料金",
-        is_deleted: null,
-      },
-      {
-        id: "TB002",
-        tenant_id: "TENANT002",
-        billing_period_start: "2025-05-01",
-        billing_period_end: "2025-05-31",
-        billing_type: "USAGE",
-        plan_id: "PLAN_ENTERPRISE",
-        plan_name: "エンタープライズプラン",
-        base_amount: 100000.0,
-        usage_amount: 45000.0,
-        additional_amount: 0.0,
-        discount_amount: 10000.0,
-        subtotal_amount: 135000.0,
-        tax_rate: 0.1,
-        tax_amount: 13500.0,
-        total_amount: 148500.0,
-        currency_code: "JPY",
-        usage_details: "{"users": 100, "storage_gb": 500, "api_calls": 200000, "premium_features": true}",
-        billing_status: "INVOICED",
-        invoice_number: "INV-2025-05-002",
-        invoice_date: "2025-06-01",
-        due_date: "2025-06-30",
-        paid_date: null,
-        payment_method: "BANK_TRANSFER",
-        notes: "5月分従量課金（大容量利用）",
-        is_deleted: null,
-      },
-    ],
-  })
-  
-  // MST_CareerPlan_sample_data.sql
-  console.log('📊 MST_CareerPlanデータを投入中...')
-  await prisma.careerPlan.createMany({
-    data: [
-      {
-        career_plan_id: "CP_001",
-        employee_id: "EMP000001",
-        plan_name: "シニアエンジニアへの成長プラン",
-        plan_description: "3年以内にシニアエンジニアとして技術リーダーシップを発揮できる人材になる",
-        plan_type: "MEDIUM_TERM",
-        target_position_id: "POS_003",
-        target_job_type_id: "JOB_001",
-        target_department_id: null,
-        current_level: "INTERMEDIATE",
-        target_level: "SENIOR",
-        plan_start_date: "2024-04-01",
-        plan_end_date: "2027-03-31",
-        milestone_1_date: "2024-12-31",
-        milestone_1_description: "AWS認定資格取得、チームリーダー経験",
-        milestone_2_date: "2025-12-31",
-        milestone_2_description: "大規模プロジェクトのテックリード担当",
-        milestone_3_date: "2026-12-31",
-        milestone_3_description: "後輩指導、技術選定の主導",
-        required_skills: "["Java", "Spring Boot", "AWS", "Docker", "Kubernetes", "チームマネジメント"]",
-        required_certifications: "["AWS認定ソリューションアーキテクト", "PMP"]",
-        required_experiences: "["チームリーダー経験", "大規模システム設計", "後輩指導"]",
-        development_actions: "["技術研修受講", "社外勉強会参加", "OSS貢献", "技術ブログ執筆"]",
-        training_plan: "["AWS研修", "リーダーシップ研修", "アーキテクチャ設計研修"]",
-        mentor_id: "EMP000010",
-        supervisor_id: "EMP000005",
-        plan_status: "ACTIVE",
-        progress_percentage: 25.5,
-        last_review_date: "2024-03-31",
-        next_review_date: "2024-06-30",
-        review_frequency: "QUARTERLY",
-        success_criteria: "技術力向上、チーム貢献、後輩育成実績",
-        risk_factors: "業務多忙による学習時間確保困難、技術変化への対応",
-        support_resources: "社内研修制度、書籍購入支援、外部セミナー参加費補助",
-        budget_allocated: 300000.0,
-        budget_used: 75000.0,
-        priority_level: "HIGH",
-        visibility_level: "MANAGER",
-        template_id: "TMPL_ENG_001",
-        custom_fields: "{"specialization": "バックエンド", "preferred_domain": "金融系"}",
-        notes: "本人の強い意欲と上司の全面的なサポートにより順調に進行中",
-        code: null,
-        name: null,
-        description: null,
-      },
-      {
-        career_plan_id: "CP_002",
-        employee_id: "EMP000002",
-        plan_name: "プロジェクトマネージャーへの転身プラン",
-        plan_description: "技術者からプロジェクトマネージャーへのキャリアチェンジ",
-        plan_type: "MANAGEMENT",
-        target_position_id: "POS_004",
-        target_job_type_id: "JOB_002",
-        target_department_id: null,
-        current_level: "SENIOR",
-        target_level: "MANAGER",
-        plan_start_date: "2024-01-01",
-        plan_end_date: "2025-12-31",
-        milestone_1_date: "2024-06-30",
-        milestone_1_description: "PMP資格取得、小規模プロジェクト管理経験",
-        milestone_2_date: "2024-12-31",
-        milestone_2_description: "中規模プロジェクトのサブPM担当",
-        milestone_3_date: null,
-        milestone_3_description: null,
-        required_skills: "["プロジェクト管理", "リーダーシップ", "コミュニケーション", "リスク管理"]",
-        required_certifications: "["PMP", "ITストラテジスト"]",
-        required_experiences: "["プロジェクト管理", "チームマネジメント", "ステークホルダー調整"]",
-        development_actions: "["PM研修受講", "PMI勉強会参加", "管理業務OJT"]",
-        training_plan: "["プロジェクトマネジメント基礎", "リーダーシップ研修", "交渉術研修"]",
-        mentor_id: "EMP000015",
-        supervisor_id: "EMP000008",
-        plan_status: "ACTIVE",
-        progress_percentage: 60.0,
-        last_review_date: "2024-04-30",
-        next_review_date: "2024-07-31",
-        review_frequency: "QUARTERLY",
-        success_criteria: "PMP取得、プロジェクト成功実績、チーム満足度向上",
-        risk_factors: "技術からマネジメントへの意識転換、人間関係構築",
-        support_resources: "PM研修制度、資格取得支援、メンター制度",
-        budget_allocated: 200000.0,
-        budget_used: 120000.0,
-        priority_level: "HIGH",
-        visibility_level: "DEPARTMENT",
-        template_id: "TMPL_MGR_001",
-        custom_fields: "{"management_style": "コーチング重視", "team_size_target": "10-15名"}",
-        notes: "技術的バックグラウンドを活かしたPMとして期待",
-        code: null,
-        name: null,
-        description: null,
-      },
-    ],
-  })
-  
-  // MST_CertificationRequirement_sample_data.sql
-  console.log('📊 MST_CertificationRequirementデータを投入中...')
-  await prisma.certificationRequirement.createMany({
-    data: [
-      {
-        requirement_id: "REQ_001",
-        requirement_name: "システムエンジニア必須資格要件",
-        requirement_description: "システムエンジニア職種における基本的な資格要件",
-        requirement_type: "JOB_TYPE",
-        target_job_type_id: "JOB_001",
-        target_position_id: null,
-        target_skill_grade_id: null,
-        target_department_id: null,
-        certification_id: "CERT_IPA_001",
-        requirement_level: "MANDATORY",
-        priority_order: 1,
-        alternative_certifications: "["基本情報技術者試験", "応用情報技術者試験"]",
-        minimum_experience_years: 2,
-        minimum_skill_level: "INTERMEDIATE",
-        grace_period_months: 12,
-        renewal_required: false,
-        renewal_interval_months: null,
-        exemption_conditions: "同等の実務経験5年以上、または関連する上位資格保有",
-        assessment_criteria: "資格証明書の提出、実務経験の確認",
-        business_justification: "技術的基礎知識の担保、顧客への信頼性向上",
-        compliance_requirement: false,
-        client_requirement: true,
-        internal_policy: true,
-        effective_start_date: "2024-04-01",
-        effective_end_date: null,
-        notification_timing: 90,
-        escalation_timing: 30,
-        cost_support_available: true,
-        cost_support_amount: 50000.0,
-        cost_support_conditions: "初回受験のみ、合格時に全額支給",
-        training_support_available: true,
-        recommended_training_programs: "["TRN_PROG_003", "TRN_PROG_004"]",
-        study_time_allocation: 2.0,
-        success_rate: 75.5,
-        average_study_hours: 150.0,
-        difficulty_rating: "MEDIUM",
-        active_flag: true,
-        created_by: "EMP000010",
-        approved_by: "EMP000005",
-        approval_date: "2024-03-15",
-        review_date: "2025-03-31",
-        notes: "新入社員は入社3年以内に取得必須",
-        code: null,
-        name: null,
-        description: null,
-      },
-      {
-        requirement_id: "REQ_002",
-        requirement_name: "プロジェクトマネージャー昇進要件",
-        requirement_description: "プロジェクトマネージャー役職への昇進に必要な資格要件",
-        requirement_type: "PROMOTION",
-        target_job_type_id: null,
-        target_position_id: "POS_004",
-        target_skill_grade_id: null,
-        target_department_id: null,
-        certification_id: "CERT_PMP_001",
-        requirement_level: "MANDATORY",
-        priority_order: 1,
-        alternative_certifications: "["プロジェクトマネージャ試験", "P2M資格"]",
-        minimum_experience_years: 5,
-        minimum_skill_level: "ADVANCED",
-        grace_period_months: 18,
-        renewal_required: true,
-        renewal_interval_months: 36,
-        exemption_conditions: "大規模プロジェクト成功実績3件以上",
-        assessment_criteria: "資格証明書、プロジェクト実績評価、360度評価",
-        business_justification: "プロジェクト管理能力の客観的証明、国際標準への準拠",
-        compliance_requirement: false,
-        client_requirement: true,
-        internal_policy: true,
-        effective_start_date: "2024-01-01",
-        effective_end_date: null,
-        notification_timing: 180,
-        escalation_timing: 60,
-        cost_support_available: true,
-        cost_support_amount: 100000.0,
-        cost_support_conditions: "受験料・研修費用全額支給、PDU維持費用も支援",
-        training_support_available: true,
-        recommended_training_programs: "["TRN_PROG_001", "TRN_PROG_005"]",
-        study_time_allocation: 4.0,
-        success_rate: 65.0,
-        average_study_hours: 300.0,
-        difficulty_rating: "HARD",
-        active_flag: true,
-        created_by: "EMP000015",
-        approved_by: "EMP000008",
-        approval_date: "2023-12-01",
-        review_date: "2024-12-31",
-        notes: "PMO部門配属者は優先的に取得支援",
-        code: null,
-        name: null,
-        description: null,
-      },
-    ],
-  })
-  
-  // MST_Certification_sample_data.sql
-  console.log('📊 MST_Certificationデータを投入中...')
-  await prisma.certification.createMany({
-    data: [
-      {
-        certification_code: "CERT_AWS_SAA",
-        certification_name: "AWS Certified Solutions Architect - Associate",
-        certification_name_en: "AWS Certified Solutions Architect - Associate",
-        issuer: "Amazon Web Services",
-        issuer_country: "US",
-        certification_category: "IT",
-        certification_level: "INTERMEDIATE",
-        validity_period_months: 36,
-        renewal_required: true,
-        renewal_requirements: "再認定試験の受験または上位資格の取得",
-        exam_fee: 15000,
-        exam_language: "日本語/英語",
-        exam_format: "ONLINE",
-        official_url: "https://aws.amazon.com/jp/certification/certified-solutions-architect-associate/",
-        description: "AWSクラウドでのソリューション設計・実装スキルを証明する資格",
-        skill_category_id: "SKILL_CAT_CLOUD",
-        is_recommended: true,
-        is_active: true,
-        code: null,
-        name: null,
-      },
-      {
-        certification_code: "CERT_PMP",
-        certification_name: "Project Management Professional",
-        certification_name_en: "Project Management Professional",
-        issuer: "Project Management Institute",
-        issuer_country: "US",
-        certification_category: "BUSINESS",
-        certification_level: "ADVANCED",
-        validity_period_months: 36,
-        renewal_required: true,
-        renewal_requirements: "60 PDU（Professional Development Units）の取得",
-        exam_fee: 55500,
-        exam_language: "日本語/英語",
-        exam_format: "BOTH",
-        official_url: "https://www.pmi.org/certifications/project-management-pmp",
-        description: "プロジェクトマネジメントの国際的な資格",
-        skill_category_id: "SKILL_CAT_PM",
-        is_recommended: true,
-        is_active: true,
-        code: null,
-        name: null,
-      },
-    ],
-  })
-  
-  // MST_Department_sample_data.sql
-  console.log('📊 MST_Departmentデータを投入中...')
-  await prisma.department.createMany({
-    data: [
-      {
-        department_code: "DEPT001",
-        department_name: "経営企画本部",
-        department_name_short: "経営企画",
+
+  // 2. 部署マスターデータ
+  console.log('📁 部署マスターデータを投入中...')
+  const departments = await Promise.all([
+    prisma.department.upsert({
+      where: { department_code: 'DEPT001' },
+      update: {},
+      create: {
+        department_code: 'DEPT001',
+        department_name: '開発部',
         parent_department_id: null,
         department_level: 1,
-        department_type: "HEADQUARTERS",
-        manager_id: "EMP000001",
-        deputy_manager_id: null,
-        cost_center_code: "CC001",
-        budget_amount: 50000000.0,
-        location: "本社ビル 10F",
-        phone_number: "03-1234-5678",
-        email_address: "planning@company.com",
-        establishment_date: "2020-04-01",
-        abolition_date: null,
-        department_status: "ACTIVE",
         sort_order: 1,
-        description: "会社全体の経営戦略立案・推進を担当",
-        code: null,
-        name: null,
+        code: 'DEPT001',
+        name: '開発部',
       },
-      {
-        department_code: "DEPT002",
-        department_name: "システム開発部",
-        department_name_short: "システム開発",
-        parent_department_id: "DEPT001",
-        department_level: 2,
-        department_type: "DEPARTMENT",
-        manager_id: "EMP000002",
-        deputy_manager_id: "EMP000003",
-        cost_center_code: "CC002",
-        budget_amount: 120000000.0,
-        location: "本社ビル 8F",
-        phone_number: "03-1234-5679",
-        email_address: "dev@company.com",
-        establishment_date: "2020-04-01",
-        abolition_date: null,
-        department_status: "ACTIVE",
+    }),
+    prisma.department.upsert({
+      where: { department_code: 'DEPT002' },
+      update: {},
+      create: {
+        department_code: 'DEPT002',
+        department_name: '営業部',
+        parent_department_id: null,
+        department_level: 1,
         sort_order: 2,
-        description: "社内システムの開発・保守・運用を担当",
-        code: null,
-        name: null,
+        code: 'DEPT002',
+        name: '営業部',
       },
-    ],
-  })
-  
-  // MST_EmployeeDepartment_sample_data.sql
-  console.log('📊 MST_EmployeeDepartmentデータを投入中...')
-  await prisma.employeeDepartment.createMany({
-    data: [
-      {
-        employee_id: "EMP000001",
-        department_id: "DEPT001",
-        assignment_type: "PRIMARY",
-        start_date: "2020-04-01",
-        end_date: null,
-        assignment_ratio: 100.0,
-        role_in_department: "チームリーダー",
-        reporting_manager_id: "EMP000010",
-        assignment_reason: "新卒入社時配属",
-        assignment_status: "ACTIVE",
-        approval_status: "APPROVED",
-        approved_by: "EMP000010",
-        approved_at: "2020-03-25 10:00:00",
-        code: null,
-        name: null,
-        description: null,
-      },
-      {
-        employee_id: "EMP000002",
-        department_id: "DEPT002",
-        assignment_type: "PRIMARY",
-        start_date: "2021-04-01",
-        end_date: null,
-        assignment_ratio: 80.0,
-        role_in_department: "開発担当",
-        reporting_manager_id: "EMP000011",
-        assignment_reason: "新卒入社時配属",
-        assignment_status: "ACTIVE",
-        approval_status: "APPROVED",
-        approved_by: "EMP000011",
-        approved_at: "2021-03-25 10:00:00",
-        code: null,
-        name: null,
-        description: null,
-      },
-      {
-        employee_id: "EMP000002",
-        department_id: "DEPT003",
-        assignment_type: "SECONDARY",
-        start_date: "2023-01-01",
-        end_date: null,
-        assignment_ratio: 20.0,
-        role_in_department: "プロジェクト支援",
-        reporting_manager_id: "EMP000012",
-        assignment_reason: "プロジェクト支援のため兼務",
-        assignment_status: "ACTIVE",
-        approval_status: "APPROVED",
-        approved_by: "EMP000012",
-        approved_at: "2022-12-20 14:00:00",
-        code: null,
-        name: null,
-        description: null,
-      },
-    ],
-  })
-  
-  // MST_EmployeeJobType_sample_data.sql
-  console.log('📊 MST_EmployeeJobTypeデータを投入中...')
-  await prisma.employeeJobType.createMany({
-    data: [
-      {
-        employee_job_type_id: "EJT_001",
-        employee_id: "EMP000001",
-        job_type_id: "JOB_001",
-        assignment_type: "PRIMARY",
-        assignment_ratio: 100.0,
-        effective_start_date: "2024-04-01",
-        effective_end_date: null,
-        assignment_reason: "NEW_HIRE",
-        assignment_status: "ACTIVE",
-        proficiency_level: "INTERMEDIATE",
-        target_proficiency_level: "ADVANCED",
-        target_achievement_date: "2025-03-31",
-        certification_requirements: "["基本情報技術者試験", "AWS認定"]",
-        skill_requirements: "["Java", "Spring Boot", "AWS", "Docker"]",
-        experience_requirements: "["Webアプリケーション開発", "チーム開発"]",
-        development_plan: "{"short_term": "AWS認定取得", "medium_term": "チームリーダー経験", "long_term": "アーキテクト昇格"}",
-        training_plan: "["TRN_PROG_002", "TRN_PROG_006"]",
-        mentor_id: "EMP000010",
-        supervisor_id: "EMP000005",
-        performance_rating: "GOOD",
-        last_evaluation_date: "2024-03-31",
-        next_evaluation_date: "2024-06-30",
-        evaluation_frequency: "QUARTERLY",
-        career_path: "シニアエンジニア → テックリード → アーキテクト",
-        strengths: "技術習得力、問題解決能力、チームワーク",
-        improvement_areas: "リーダーシップ、プレゼンテーション",
-        achievements: "新人研修システム開発、パフォーマンス改善20%達成",
-        goals: "AWS認定取得、チームリーダー経験積む",
-        workload_percentage: 100.0,
-        billable_flag: true,
-        cost_center: "DEV001",
-        budget_allocation: 5000000.0,
-        hourly_rate: 3500.0,
-        overtime_eligible: true,
-        remote_work_eligible: true,
-        travel_required: false,
-        security_clearance_required: false,
-        created_by: "EMP000005",
-        approved_by: "EMP000008",
-        approval_date: "2024-03-25",
-        notes: "新卒採用、高いポテンシャルを持つ",
-        code: null,
-        name: null,
-        description: null,
-      },
-      {
-        employee_job_type_id: "EJT_002",
-        employee_id: "EMP000002",
-        job_type_id: "JOB_002",
-        assignment_type: "PRIMARY",
-        assignment_ratio: 80.0,
-        effective_start_date: "2024-01-01",
-        effective_end_date: null,
-        assignment_reason: "PROMOTION",
-        assignment_status: "ACTIVE",
-        proficiency_level: "ADVANCED",
-        target_proficiency_level: "EXPERT",
-        target_achievement_date: "2024-12-31",
-        certification_requirements: "["PMP", "ITストラテジスト"]",
-        skill_requirements: "["プロジェクト管理", "リーダーシップ", "ステークホルダー管理"]",
-        experience_requirements: "["大規模プロジェクト管理", "チームマネジメント"]",
-        development_plan: "{"short_term": "PMP取得", "medium_term": "大規模PM経験", "long_term": "PMO責任者"}",
-        training_plan: "["TRN_PROG_001", "TRN_PROG_007"]",
-        mentor_id: "EMP000015",
-        supervisor_id: "EMP000008",
-        performance_rating: "EXCELLENT",
-        last_evaluation_date: "2024-04-30",
-        next_evaluation_date: "2024-07-31",
-        evaluation_frequency: "QUARTERLY",
-        career_path: "プロジェクトマネージャー → シニアPM → PMO責任者",
-        strengths: "プロジェクト管理、コミュニケーション、問題解決",
-        improvement_areas: "戦略立案、予算管理",
-        achievements: "3つの大規模プロジェクト成功、チーム満足度向上",
-        goals: "PMP取得、PMO体制構築",
-        workload_percentage: 80.0,
-        billable_flag: true,
-        cost_center: "PMO001",
-        budget_allocation: 8000000.0,
-        hourly_rate: 5000.0,
-        overtime_eligible: false,
-        remote_work_eligible: true,
-        travel_required: true,
-        security_clearance_required: false,
-        created_by: "EMP000008",
-        approved_by: "EMP000001",
-        approval_date: "2023-12-15",
-        notes: "技術者からPMへの転身成功例",
-        code: null,
-        name: null,
-        description: null,
-      },
-    ],
-  })
-  
-  // MST_EmployeePosition_sample_data.sql
-  console.log('📊 MST_EmployeePositionデータを投入中...')
-  await prisma.employeePosition.createMany({
-    data: [
-      {
-        employee_id: "EMP000001",
-        position_id: "POS001",
-        appointment_type: "PRIMARY",
-        start_date: "2020-04-01",
-        end_date: null,
-        appointment_reason: "新卒入社時任命",
-        responsibility_scope: "チーム運営、メンバー指導、プロジェクト管理",
-        authority_level: 5,
-        salary_grade: "G5",
-        appointment_status: "ACTIVE",
-        approval_status: "APPROVED",
-        approved_by: "EMP000010",
-        approved_at: "2020-03-25 10:00:00",
-        performance_target: "チーム生産性20%向上、メンバー育成2名",
-        delegation_authority: "{"budget_approval": 1000000, "hiring_authority": true, "performance_evaluation": true}",
-        code: null,
-        name: null,
-        description: null,
-      },
-      {
-        employee_id: "EMP000002",
-        position_id: "POS002",
-        appointment_type: "PRIMARY",
-        start_date: "2021-04-01",
-        end_date: "2023-03-31",
-        appointment_reason: "新卒入社時任命",
-        responsibility_scope: "システム開発、技術調査",
-        authority_level: 3,
-        salary_grade: "G3",
-        appointment_status: "INACTIVE",
-        approval_status: "APPROVED",
-        approved_by: "EMP000011",
-        approved_at: "2021-03-25 10:00:00",
-        performance_target: "開発効率向上、技術スキル習得",
-        delegation_authority: "{"code_review": true, "technical_decision": false}",
-        code: null,
-        name: null,
-        description: null,
-      },
-      {
-        employee_id: "EMP000002",
-        position_id: "POS003",
-        appointment_type: "PRIMARY",
-        start_date: "2023-04-01",
-        end_date: null,
-        appointment_reason: "昇進による任命",
-        responsibility_scope: "シニア開発者、技術指導、アーキテクチャ設計",
-        authority_level: 4,
-        salary_grade: "G4",
-        appointment_status: "ACTIVE",
-        approval_status: "APPROVED",
-        approved_by: "EMP000011",
-        approved_at: "2023-03-20 14:00:00",
-        performance_target: "技術品質向上、後輩指導3名",
-        delegation_authority: "{"technical_decision": true, "architecture_review": true}",
-        code: null,
-        name: null,
-        description: null,
-      },
-    ],
-  })
-  
-  // MST_Employee_sample_data.sql
-  console.log('📊 MST_Employeeデータを投入中...')
-  await prisma.employee.createMany({
-    data: [
-      {
-        employee_code: "EMP000001",
-        full_name: "山田太郎",
-        full_name_kana: "ヤマダタロウ",
-        email: "yamada.taro@company.com",
-        phone: "090-1234-5678",
-        hire_date: "2020-04-01",
-        birth_date: "1990-01-15",
-        gender: "M",
-        department_id: "DEPT001",
-        position_id: "POS001",
-        job_type_id: "JOB001",
-        employment_status: "FULL_TIME",
-        manager_id: null,
-        employee_status: "ACTIVE",
-        code: null,
-        name: null,
-        description: null,
-      },
-      {
-        employee_code: "EMP000002",
-        full_name: "佐藤花子",
-        full_name_kana: "サトウハナコ",
-        email: "sato.hanako@company.com",
-        phone: "090-2345-6789",
-        hire_date: "2021-04-01",
-        birth_date: "1992-03-20",
-        gender: "F",
-        department_id: "DEPT002",
-        position_id: "POS002",
-        job_type_id: "JOB002",
-        employment_status: "FULL_TIME",
-        manager_id: "EMP000001",
-        employee_status: "ACTIVE",
-        code: null,
-        name: null,
-        description: null,
-      },
-    ],
-  })
-  
-  // MST_JobTypeSkillGrade_sample_data.sql
-  console.log('📊 MST_JobTypeSkillGradeデータを投入中...')
-  await prisma.jobTypeSkillGrade.createMany({
-    data: [
-      {
-        job_type_id: "JOB001",
-        skill_grade_id: "GRADE001",
-        grade_requirement_type: "MINIMUM",
-        required_experience_years: 0.0,
-        promotion_criteria: "基礎研修修了、OJT完了、基本業務遂行能力",
-        salary_range_min: 3000000,
-        salary_range_max: 4000000,
-        performance_expectations: "指導の下での基本業務遂行、学習意欲の継続",
-        leadership_requirements: "チームワーク、積極的な学習姿勢",
-        technical_depth: 2,
-        business_impact: 2,
-        team_size_expectation: 0,
-        certification_requirements: "["基本情報技術者試験"]",
-        grade_status: "ACTIVE",
-        effective_date: "2025-01-01",
-        expiry_date: null,
-        next_grade_path: "["GRADE002"]",
-        evaluation_frequency: "SEMI_ANNUAL",
-        code: null,
-        name: null,
-        description: null,
-      },
-      {
-        job_type_id: "JOB001",
-        skill_grade_id: "GRADE002",
-        grade_requirement_type: "STANDARD",
-        required_experience_years: 2.0,
-        promotion_criteria: "独立した業務遂行、技術スキル向上、プロジェクト貢献",
-        salary_range_min: 4000000,
-        salary_range_max: 5500000,
-        performance_expectations: "独立した業務遂行、品質向上への貢献、後輩指導",
-        leadership_requirements: "後輩指導、技術的リーダーシップ",
-        technical_depth: 4,
-        business_impact: 4,
-        team_size_expectation: 2,
-        certification_requirements: "["応用情報技術者試験", "専門資格1つ以上"]",
-        grade_status: "ACTIVE",
-        effective_date: "2025-01-01",
-        expiry_date: null,
-        next_grade_path: "["GRADE003"]",
-        evaluation_frequency: "ANNUAL",
-        code: null,
-        name: null,
-        description: null,
-      },
-      {
-        job_type_id: "JOB002",
-        skill_grade_id: "GRADE003",
-        grade_requirement_type: "ADVANCED",
-        required_experience_years: 5.0,
-        promotion_criteria: "チーム運営実績、技術的専門性、事業貢献度",
-        salary_range_min: 6000000,
-        salary_range_max: 8000000,
-        performance_expectations: "チーム運営、技術戦略立案、事業成果創出",
-        leadership_requirements: "チームマネジメント、技術戦略、人材育成",
-        technical_depth: 7,
-        business_impact: 7,
-        team_size_expectation: 5,
-        certification_requirements: "["高度情報技術者試験", "マネジメント資格", "専門資格2つ以上"]",
-        grade_status: "ACTIVE",
-        effective_date: "2025-01-01",
-        expiry_date: null,
-        next_grade_path: "["GRADE004", "GRADE005"]",
-        evaluation_frequency: "ANNUAL",
-        code: null,
-        name: null,
-        description: null,
-      },
-    ],
-  })
-  
-  // MST_JobTypeSkill_sample_data.sql
-  console.log('📊 MST_JobTypeSkillデータを投入中...')
-  await prisma.jobTypeSkill.createMany({
-    data: [
-      {
-        job_type_id: "JOB001",
-        skill_item_id: "SKILL001",
-        required_level: 4,
-        skill_priority: "CRITICAL",
-        skill_category: "TECHNICAL",
-        experience_years: 3.0,
-        certification_required: true,
-        skill_weight: 25.0,
-        evaluation_criteria: "実務プロジェクトでの設計・実装経験、コードレビュー能力",
-        learning_path: "基礎研修→実践プロジェクト→上級研修→資格取得",
-        skill_status: "ACTIVE",
-        effective_date: "2025-01-01",
-        expiry_date: null,
-        alternative_skills: "["SKILL002", "SKILL003"]",
-        prerequisite_skills: "["SKILL010", "SKILL011"]",
-        code: null,
-        name: null,
-        description: null,
-      },
-      {
-        job_type_id: "JOB001",
-        skill_item_id: "SKILL002",
-        required_level: 3,
-        skill_priority: "HIGH",
-        skill_category: "BUSINESS",
-        experience_years: 2.0,
-        certification_required: false,
-        skill_weight: 20.0,
-        evaluation_criteria: "業務要件の理解度、顧客とのコミュニケーション能力",
-        learning_path: "業務知識研修→OJT→実践経験",
-        skill_status: "ACTIVE",
-        effective_date: "2025-01-01",
-        expiry_date: null,
-        alternative_skills: "["SKILL004"]",
-        prerequisite_skills: "["SKILL012"]",
-        code: null,
-        name: null,
-        description: null,
-      },
-      {
-        job_type_id: "JOB002",
-        skill_item_id: "SKILL003",
-        required_level: 5,
-        skill_priority: "CRITICAL",
-        skill_category: "MANAGEMENT",
-        experience_years: 5.0,
-        certification_required: true,
-        skill_weight: 30.0,
-        evaluation_criteria: "チーム運営実績、プロジェクト成功率、メンバー育成実績",
-        learning_path: "リーダーシップ研修→実践経験→管理職研修→資格取得",
-        skill_status: "ACTIVE",
-        effective_date: "2025-01-01",
-        expiry_date: null,
-        alternative_skills: null,
-        prerequisite_skills: "["SKILL001", "SKILL002"]",
-        code: null,
-        name: null,
-        description: null,
-      },
-    ],
-  })
-  
-  // MST_JobType_sample_data.sql
-  console.log('📊 MST_JobTypeデータを投入中...')
-  await prisma.jobType.createMany({
-    data: [
-      {
-        job_type_code: "SE",
-        job_type_name: "システムエンジニア",
-        job_type_name_en: "Systems Engineer",
-        job_category: "ENGINEERING",
-        job_level: "SENIOR",
-        description: "システムの設計・開発・テストを担当するエンジニア",
-        required_experience_years: 3,
-        salary_grade_min: 3,
-        salary_grade_max: 6,
-        career_path: "SE → シニアSE → テックリード → エンジニアリングマネージャー",
-        required_certifications: "["基本情報技術者", "応用情報技術者"]",
-        required_skills: "["Java", "SQL", "システム設計", "要件定義"]",
-        department_affinity: "["開発部", "システム部"]",
-        remote_work_eligible: true,
-        travel_frequency: "LOW",
-        sort_order: 1,
-        is_active: true,
-        code: null,
-        name: null,
-      },
-      {
-        job_type_code: "PM",
-        job_type_name: "プロジェクトマネージャー",
-        job_type_name_en: "Project Manager",
-        job_category: "MANAGEMENT",
-        job_level: "MANAGER",
-        description: "プロジェクトの計画・実行・管理を統括する責任者",
-        required_experience_years: 5,
-        salary_grade_min: 5,
-        salary_grade_max: 8,
-        career_path: "SE → リーダー → PM → 部門マネージャー",
-        required_certifications: "["PMP", "プロジェクトマネージャ試験"]",
-        required_skills: "["プロジェクト管理", "リーダーシップ", "コミュニケーション", "リスク管理"]",
-        department_affinity: "["開発部", "PMO"]",
-        remote_work_eligible: true,
-        travel_frequency: "MEDIUM",
-        sort_order: 2,
-        is_active: true,
-        code: null,
-        name: null,
-      },
-      {
-        job_type_code: "QA",
-        job_type_name: "品質保証エンジニア",
-        job_type_name_en: "Quality Assurance Engineer",
-        job_category: "ENGINEERING",
-        job_level: "SENIOR",
-        description: "ソフトウェアの品質保証・テスト設計・実行を担当",
-        required_experience_years: 2,
-        salary_grade_min: 3,
-        salary_grade_max: 6,
-        career_path: "QA → シニアQA → QAリード → QAマネージャー",
-        required_certifications: "["JSTQB", "ソフトウェア品質技術者資格"]",
-        required_skills: "["テスト設計", "自動化テスト", "品質管理", "バグ分析"]",
-        department_affinity: "["品質保証部", "開発部"]",
-        remote_work_eligible: true,
-        travel_frequency: "NONE",
+    }),
+    prisma.department.upsert({
+      where: { department_code: 'DEPT003' },
+      update: {},
+      create: {
+        department_code: 'DEPT003',
+        department_name: '管理部',
+        parent_department_id: null,
+        department_level: 1,
         sort_order: 3,
-        is_active: true,
-        code: null,
-        name: null,
+        code: 'DEPT003',
+        name: '管理部',
       },
-    ],
-  })
-  
-  // MST_NotificationSettings_sample_data.sql
-  console.log('📊 MST_NotificationSettingsデータを投入中...')
-  await prisma.notificationSettings.createMany({
-    data: [
-      {
-        id: "NS001",
-        tenant_id: "TENANT001",
-        setting_key: "skill_update_notification",
-        setting_name: "スキル更新通知",
-        notification_type: "EMAIL",
-        target_audience: "MANAGER",
-        trigger_event: "skill_registered",
-        frequency_type: "IMMEDIATE",
-        frequency_value: null,
-        template_id: "NT001",
-        channel_config: "{"smtp_server": "smtp.company.com", "from_address": "noreply@company.com"}",
-        is_enabled: true,
-        priority_level: "MEDIUM",
-        code: null,
-        name: null,
-        description: null,
-      },
-      {
-        id: "NS002",
-        tenant_id: "TENANT001",
-        setting_key: "goal_deadline_reminder",
-        setting_name: "目標期限リマインダー",
-        notification_type: "SLACK",
-        target_audience: "EMPLOYEE",
-        trigger_event: "goal_deadline_approaching",
-        frequency_type: "DAILY",
-        frequency_value: 9,
-        template_id: "NT002",
-        channel_config: "{"webhook_url": "https://hooks.slack.com/services/xxx", "channel": "#notifications"}",
-        is_enabled: true,
-        priority_level: "HIGH",
-        code: null,
-        name: null,
-        description: null,
-      },
-    ],
-  })
-  
-  // MST_NotificationTemplate_sample_data.sql
-  console.log('📊 MST_NotificationTemplateデータを投入中...')
-  await prisma.notificationTemplate.createMany({
-    data: [
-      {
-        id: "NT001",
-        tenant_id: "TENANT001",
-        template_key: "skill_update_notification",
-        template_name: "スキル更新通知テンプレート",
-        notification_type: "EMAIL",
-        language_code: "ja",
-        subject_template: "【スキル更新】{{employee_name}}さんのスキル情報が更新されました",
-        body_template: "{{employee_name}}さん
-  
-  以下のスキル情報が更新されました。
-  
-  スキル名: {{skill_name}}
-  更新日時: {{updated_at}}
-  更新者: {{updated_by}}
-  
-  詳細は以下のリンクからご確認ください。
-  {{skill_detail_url}}
-  
-  ※このメールは自動送信されています。
-  ",
-        format_type: "PLAIN",
-        parameters: "{"employee_name": "社員名", "skill_name": "スキル名", "updated_at": "更新日時", "updated_by": "更新者", "skill_detail_url": "詳細URL"}",
-        sample_data: "{"employee_name": "山田太郎", "skill_name": "Java", "updated_at": "2025-06-01 10:30:00", "updated_by": "佐藤花子", "skill_detail_url": "https://system.company.com/skills/123"}",
-        is_default: true,
-        is_active: true,
-        version: "1.0.0",
-        code: null,
-        name: null,
-        description: null,
-      },
-      {
-        id: "NT002",
-        tenant_id: "TENANT001",
-        template_key: "goal_deadline_reminder",
-        template_name: "目標期限リマインダーテンプレート",
-        notification_type: "SLACK",
-        language_code: "ja",
-        subject_template: null,
-        body_template: ":warning: *目標期限のお知らせ* :warning:
-  
-  {{employee_name}}さんの目標「{{goal_title}}」の期限が近づいています。
-  
-  • 期限: {{deadline_date}}
-  • 残り日数: {{remaining_days}}日
-  • 進捗率: {{progress_rate}}%
-  
-  <{{goal_detail_url}}|詳細を確認する>
-  ",
-        format_type: "MARKDOWN",
-        parameters: "{"employee_name": "社員名", "goal_title": "目標タイトル", "deadline_date": "期限日", "remaining_days": "残り日数", "progress_rate": "進捗率", "goal_detail_url": "詳細URL"}",
-        sample_data: "{"employee_name": "山田太郎", "goal_title": "Java認定資格取得", "deadline_date": "2025-06-30", "remaining_days": "29", "progress_rate": "75", "goal_detail_url": "https://system.company.com/goals/456"}",
-        is_default: true,
-        is_active: true,
-        version: "1.0.0",
-        code: null,
-        name: null,
-        description: null,
-      },
-    ],
-  })
-  
-  // MST_Permission_sample_data.sql
-  console.log('📊 MST_Permissionデータを投入中...')
-  await prisma.permission.createMany({
-    data: [
-      {
-        permission_code: "PERM_USER_READ",
-        permission_name: "ユーザー情報参照",
-        permission_name_short: "ユーザー参照",
-        permission_category: "DATA",
-        resource_type: "USER",
-        action_type: "READ",
-        scope_level: "TENANT",
-        parent_permission_id: null,
-        is_system_permission: true,
-        requires_conditions: false,
-        condition_expression: null,
-        risk_level: 1,
-        requires_approval: false,
-        audit_required: true,
-        permission_status: "ACTIVE",
-        effective_from: "2025-01-01",
-        effective_to: null,
+    }),
+  ])
+
+  // 3. 役職マスターデータ
+  console.log('👔 役職マスターデータを投入中...')
+  const positions = await Promise.all([
+    prisma.position.upsert({
+      where: { position_code: 'POS001' },
+      update: {},
+      create: {
+        position_code: 'POS001',
+        position_name: '部長',
+        position_level: 4,
         sort_order: 1,
-        description: "ユーザー情報の参照権限",
-        code: null,
-        name: null,
+        code: 'POS001',
+        name: '部長',
       },
-      {
-        permission_code: "PERM_USER_UPDATE",
-        permission_name: "ユーザー情報更新",
-        permission_name_short: "ユーザー更新",
-        permission_category: "DATA",
-        resource_type: "USER",
-        action_type: "UPDATE",
-        scope_level: "DEPARTMENT",
-        parent_permission_id: null,
-        is_system_permission: true,
-        requires_conditions: true,
-        condition_expression: "department_id = :user_department_id",
-        risk_level: 2,
-        requires_approval: false,
-        audit_required: true,
-        permission_status: "ACTIVE",
-        effective_from: "2025-01-01",
-        effective_to: null,
-        sort_order: 2,
-        description: "ユーザー情報の更新権限（同一部署のみ）",
-        code: null,
-        name: null,
-      },
-      {
-        permission_code: "PERM_SYSTEM_ADMIN",
-        permission_name: "システム管理",
-        permission_name_short: "システム管理",
-        permission_category: "SYSTEM",
-        resource_type: "SYSTEM",
-        action_type: "EXECUTE",
-        scope_level: "GLOBAL",
-        parent_permission_id: null,
-        is_system_permission: true,
-        requires_conditions: false,
-        condition_expression: null,
-        risk_level: 4,
-        requires_approval: true,
-        audit_required: true,
-        permission_status: "ACTIVE",
-        effective_from: "2025-01-01",
-        effective_to: null,
-        sort_order: 100,
-        description: "システム全体の管理権限",
-        code: null,
-        name: null,
-      },
-    ],
-  })
-  
-  // MST_Position_sample_data.sql
-  console.log('📊 MST_Positionデータを投入中...')
-  await prisma.position.createMany({
-    data: [
-      {
-        position_code: "POS001",
-        position_name: "代表取締役社長",
-        position_name_short: "社長",
-        position_level: 1,
-        position_rank: 1,
-        position_category: "EXECUTIVE",
-        authority_level: 10,
-        approval_limit: 999999999.99,
-        salary_grade: "E1",
-        allowance_amount: 500000.0,
-        is_management: true,
-        is_executive: true,
-        requires_approval: true,
-        can_hire: true,
-        can_evaluate: true,
-        position_status: "ACTIVE",
-        sort_order: 1,
-        description: "会社の最高責任者として経営全般を統括",
-        code: null,
-        name: null,
-      },
-      {
-        position_code: "POS002",
-        position_name: "取締役",
-        position_name_short: "取締役",
-        position_level: 2,
-        position_rank: 1,
-        position_category: "EXECUTIVE",
-        authority_level: 9,
-        approval_limit: 100000000.0,
-        salary_grade: "E2",
-        allowance_amount: 300000.0,
-        is_management: true,
-        is_executive: true,
-        requires_approval: true,
-        can_hire: true,
-        can_evaluate: true,
-        position_status: "ACTIVE",
-        sort_order: 2,
-        description: "取締役会メンバーとして経営方針決定に参画",
-        code: null,
-        name: null,
-      },
-      {
-        position_code: "POS003",
-        position_name: "部長",
-        position_name_short: "部長",
+    }),
+    prisma.position.upsert({
+      where: { position_code: 'POS002' },
+      update: {},
+      create: {
+        position_code: 'POS002',
+        position_name: '課長',
         position_level: 3,
-        position_rank: 1,
-        position_category: "MANAGER",
-        authority_level: 7,
-        approval_limit: 10000000.0,
-        salary_grade: "M1",
-        allowance_amount: 100000.0,
-        is_management: true,
-        is_executive: false,
-        requires_approval: true,
-        can_hire: true,
-        can_evaluate: true,
-        position_status: "ACTIVE",
-        sort_order: 3,
-        description: "部門の責任者として業務全般を管理",
-        code: null,
-        name: null,
-      },
-    ],
-  })
-  
-  // MST_ReportTemplate_sample_data.sql
-  console.log('📊 MST_ReportTemplateデータを投入中...')
-  await prisma.reportTemplate.createMany({
-    data: [
-      {
-        id: "RT001",
-        tenant_id: "TENANT001",
-        template_key: "skill_summary_report",
-        template_name: "スキルサマリーレポート",
-        report_category: "SKILL",
-        output_format: "PDF",
-        language_code: "ja",
-        template_content: "<!DOCTYPE html>
-  <html>
-  <head>
-      <meta charset="UTF-8">
-      <title>{{report_title}}</title>
-  </head>
-  <body>
-      <h1>{{employee_name}}さんのスキルサマリー</h1>
-      <div class="summary-section">
-          <h2>保有スキル一覧</h2>
-          {{#skills}}
-          <div class="skill-item">
-              <span class="skill-name">{{skill_name}}</span>
-              <span class="skill-level">レベル: {{skill_level}}</span>
-          </div>
-          {{/skills}}
-      </div>
-  </body>
-  </html>
-  ",
-        style_sheet: "body { font-family: ''Noto Sans JP'', sans-serif; }
-  .skill-item { margin: 10px 0; padding: 5px; border-bottom: 1px solid #ccc; }
-  .skill-name { font-weight: bold; }
-  .skill-level { color: #666; }
-  ",
-        parameters_schema: "{"type": "object", "properties": {"employee_id": {"type": "string"}, "report_date": {"type": "string", "format": "date"}}}",
-        data_source_config: "{"tables": ["MST_Employee", "TRN_EmployeeSkill", "MST_Skill"], "joins": ["employee_skills", "skill_details"]}",
-        page_settings: "{"size": "A4", "orientation": "portrait", "margin": {"top": "20mm", "bottom": "20mm", "left": "15mm", "right": "15mm"}}",
-        header_template: "<div style="text-align: center; font-size: 12px;">{{company_name}} - スキル管理システム</div>",
-        footer_template: "<div style="text-align: center; font-size: 10px;">出力日時: {{generated_at}} - ページ {{page_number}}</div>",
-        is_default: true,
-        is_active: true,
-        version: "1.0.0",
-        preview_image_url: "/assets/templates/skill_summary_preview.png",
-        code: null,
-        name: null,
-        description: null,
-      },
-      {
-        id: "RT002",
-        tenant_id: "TENANT001",
-        template_key: "goal_progress_report",
-        template_name: "目標進捗レポート",
-        report_category: "GOAL",
-        output_format: "EXCEL",
-        language_code: "ja",
-        template_content: "<workbook>
-      <worksheet name="目標進捗">
-          <row>
-              <cell>社員名</cell>
-              <cell>目標タイトル</cell>
-              <cell>進捗率</cell>
-              <cell>期限</cell>
-              <cell>状態</cell>
-          </row>
-          {{#goals}}
-          <row>
-              <cell>{{employee_name}}</cell>
-              <cell>{{goal_title}}</cell>
-              <cell>{{progress_rate}}%</cell>
-              <cell>{{deadline}}</cell>
-              <cell>{{status}}</cell>
-          </row>
-          {{/goals}}
-      </worksheet>
-  </workbook>
-  ",
-        style_sheet: null,
-        parameters_schema: "{"type": "object", "properties": {"department_id": {"type": "string"}, "period_start": {"type": "string", "format": "date"}, "period_end": {"type": "string", "format": "date"}}}",
-        data_source_config: "{"tables": ["MST_Employee", "TRN_Goal"], "joins": ["employee_goals"]}",
-        page_settings: "{"orientation": "landscape"}",
-        header_template: null,
-        footer_template: null,
-        is_default: true,
-        is_active: true,
-        version: "1.0.0",
-        preview_image_url: "/assets/templates/goal_progress_preview.png",
-        code: null,
-        name: null,
-        description: null,
-      },
-    ],
-  })
-  
-  // MST_Role_sample_data.sql
-  console.log('📊 MST_Roleデータを投入中...')
-  await prisma.role.createMany({
-    data: [
-      {
-        role_code: "ROLE001",
-        role_name: "システム管理者",
-        role_name_short: "システム管理者",
-        role_category: "SYSTEM",
-        role_level: 1,
-        parent_role_id: null,
-        is_system_role: true,
-        is_tenant_specific: false,
-        max_users: 5,
-        role_priority: 1,
-        auto_assign_conditions: null,
-        role_status: "ACTIVE",
-        effective_from: "2025-01-01",
-        effective_to: null,
-        sort_order: 1,
-        description: "システム全体の管理権限を持つ最上位ロール",
-        code: null,
-        name: null,
-      },
-      {
-        role_code: "ROLE002",
-        role_name: "テナント管理者",
-        role_name_short: "テナント管理者",
-        role_category: "TENANT",
-        role_level: 2,
-        parent_role_id: null,
-        is_system_role: true,
-        is_tenant_specific: true,
-        max_users: 10,
-        role_priority: 2,
-        auto_assign_conditions: null,
-        role_status: "ACTIVE",
-        effective_from: "2025-01-01",
-        effective_to: null,
         sort_order: 2,
-        description: "テナント内の管理権限を持つロール",
-        code: null,
-        name: null,
+        code: 'POS002',
+        name: '課長',
       },
-      {
-        role_code: "ROLE003",
-        role_name: "一般ユーザー",
-        role_name_short: "一般ユーザー",
-        role_category: "BUSINESS",
-        role_level: 3,
-        parent_role_id: null,
-        is_system_role: true,
-        is_tenant_specific: false,
-        max_users: null,
-        role_priority: 10,
-        auto_assign_conditions: "{"default": true}",
-        role_status: "ACTIVE",
-        effective_from: "2025-01-01",
-        effective_to: null,
-        sort_order: 10,
-        description: "基本的な業務機能を利用できるロール",
-        code: null,
-        name: null,
-      },
-    ],
-  })
-  
-  // MST_SkillCategory_sample_data.sql
-  console.log('📊 MST_SkillCategoryデータを投入中...')
-  await prisma.skillCategory.createMany({
-    data: [
-      {
-        category_code: "CAT001",
-        category_name: "プログラミング言語",
-        category_name_short: "プログラミング",
-        category_name_en: "Programming Languages",
-        category_type: "TECHNICAL",
-        parent_category_id: null,
-        category_level: 1,
-        category_path: "/プログラミング言語",
-        is_system_category: true,
-        is_leaf_category: false,
-        skill_count: 25,
-        evaluation_method: "LEVEL",
-        max_level: 5,
-        icon_url: "/icons/programming.svg",
-        color_code: "#007ACC",
-        display_order: 1,
-        is_popular: true,
-        category_status: "ACTIVE",
-        effective_from: "2025-01-01",
-        effective_to: null,
-        description: "各種プログラミング言語のスキル",
-        code: null,
-        name: null,
-      },
-      {
-        category_code: "CAT002",
-        category_name: "Java",
-        category_name_short: "Java",
-        category_name_en: "Java",
-        category_type: "TECHNICAL",
-        parent_category_id: "CAT001",
-        category_level: 2,
-        category_path: "/プログラミング言語/Java",
-        is_system_category: true,
-        is_leaf_category: true,
-        skill_count: 8,
-        evaluation_method: "LEVEL",
-        max_level: 5,
-        icon_url: "/icons/java.svg",
-        color_code: "#ED8B00",
-        display_order: 1,
-        is_popular: true,
-        category_status: "ACTIVE",
-        effective_from: "2025-01-01",
-        effective_to: null,
-        description: "Java言語に関するスキル",
-        code: null,
-        name: null,
-      },
-      {
-        category_code: "CAT003",
-        category_name: "コミュニケーション",
-        category_name_short: "コミュニケーション",
-        category_name_en: "Communication",
-        category_type: "SOFT",
-        parent_category_id: null,
-        category_level: 1,
-        category_path: "/コミュニケーション",
-        is_system_category: true,
-        is_leaf_category: true,
-        skill_count: 12,
-        evaluation_method: "LEVEL",
-        max_level: 4,
-        icon_url: "/icons/communication.svg",
-        color_code: "#28A745",
-        display_order: 10,
-        is_popular: true,
-        category_status: "ACTIVE",
-        effective_from: "2025-01-01",
-        effective_to: null,
-        description: "コミュニケーション能力に関するスキル",
-        code: null,
-        name: null,
-      },
-    ],
-  })
-  
-  // MST_SkillGradeRequirement_sample_data.sql
-  console.log('📊 MST_SkillGradeRequirementデータを投入中...')
-  await prisma.skillGradeRequirement.createMany({
-    data: [
-      {
-        skill_grade_id: "GRADE001",
-        requirement_category: "TECHNICAL",
-        requirement_name: "プログラミング基礎",
-        requirement_description: "基本的なプログラミング言語の理解と簡単なプログラムの作成能力",
-        evaluation_criteria: "指定された仕様に基づく簡単なプログラムの作成、基本的なアルゴリズムの理解",
-        proficiency_level: 2,
-        weight_percentage: 30.0,
-        minimum_score: 70.0,
-        evidence_requirements: "作成したプログラムのソースコード、動作確認結果",
-        learning_resources: "["プログラミング入門書", "オンライン学習サイト", "基礎研修"]",
-        prerequisite_requirements: "["コンピュータ基礎知識"]",
-        assessment_method: "PROJECT",
-        assessment_frequency: "SEMI_ANNUAL",
-        validity_period: 24,
-        certification_mapping: "["基本情報技術者試験"]",
-        requirement_status: "ACTIVE",
-        effective_date: "2025-01-01",
-        expiry_date: null,
-        revision_notes: "初版作成",
-        code: null,
-        name: null,
-        description: null,
-      },
-      {
-        skill_grade_id: "GRADE001",
-        requirement_category: "BUSINESS",
-        requirement_name: "業務理解",
-        requirement_description: "担当業務の基本的な理解と顧客要件の把握能力",
-        evaluation_criteria: "業務フローの説明、顧客要件の整理と文書化",
-        proficiency_level: 2,
-        weight_percentage: 25.0,
-        minimum_score: 75.0,
-        evidence_requirements: "業務分析レポート、要件定義書",
-        learning_resources: "["業務知識研修", "業界動向資料", "先輩社員からのOJT"]",
-        prerequisite_requirements: null,
-        assessment_method: "PORTFOLIO",
-        assessment_frequency: "ANNUAL",
-        validity_period: 12,
-        certification_mapping: null,
-        requirement_status: "ACTIVE",
-        effective_date: "2025-01-01",
-        expiry_date: null,
-        revision_notes: "初版作成",
-        code: null,
-        name: null,
-        description: null,
-      },
-      {
-        skill_grade_id: "GRADE003",
-        requirement_category: "LEADERSHIP",
-        requirement_name: "チームマネジメント",
-        requirement_description: "チームの運営管理と成果創出のためのリーダーシップ能力",
-        evaluation_criteria: "チーム運営実績、メンバー育成成果、プロジェクト成功率",
-        proficiency_level: 4,
-        weight_percentage: 35.0,
-        minimum_score: 80.0,
-        evidence_requirements: "チーム運営レポート、メンバー評価、プロジェクト成果物",
-        learning_resources: "["リーダーシップ研修", "マネジメント書籍", "外部セミナー"]",
-        prerequisite_requirements: "["チームリーダー経験", "プロジェクト管理経験"]",
-        assessment_method: "PEER_REVIEW",
-        assessment_frequency: "ANNUAL",
-        validity_period: 36,
-        certification_mapping: "["PMP", "プロジェクトマネージャ試験"]",
-        requirement_status: "ACTIVE",
-        effective_date: "2025-01-01",
-        expiry_date: null,
-        revision_notes: "初版作成",
-        code: null,
-        name: null,
-        description: null,
-      },
-    ],
-  })
-  
-  // MST_SkillGrade_sample_data.sql
-  console.log('📊 MST_SkillGradeデータを投入中...')
-  await prisma.skillGrade.createMany({
-    data: [
-      {
-        grade_code: "BEGINNER",
-        grade_name: "初級",
-        grade_name_short: "初級",
-        grade_level: 1,
-        description: "基本的な知識を持ち、指導の下で業務を遂行できるレベル",
-        evaluation_criteria: "基本概念の理解、簡単なタスクの実行、指導者のサポートが必要",
-        required_experience_months: 6,
-        skill_indicators: "["基本知識", "指導下での作業", "学習意欲"]",
-        competency_requirements: "["基礎理論の理解", "基本操作の習得"]",
-        certification_requirements: "[]",
-        project_complexity: "SIMPLE",
-        mentoring_capability: false,
-        leadership_level: "NONE",
-        salary_impact_factor: 1.0,
-        promotion_eligibility: false,
-        color_code: "#90EE90",
-        sort_order: 1,
-        is_active: true,
-        code: null,
-        name: null,
-      },
-      {
-        grade_code: "INTERMEDIATE",
-        grade_name: "中級",
-        grade_name_short: "中級",
-        grade_level: 2,
-        description: "一般的な業務を独立して遂行でき、部分的に他者を指導できるレベル",
-        evaluation_criteria: "独立した作業遂行、問題解決能力、基本的な指導スキル",
-        required_experience_months: 18,
-        skill_indicators: "["独立作業", "問題解決", "基本指導"]",
-        competency_requirements: "["実践的スキル", "問題分析能力", "コミュニケーション能力"]",
-        certification_requirements: "["基本情報技術者"]",
-        project_complexity: "MODERATE",
-        mentoring_capability: true,
-        leadership_level: "TEAM",
-        salary_impact_factor: 1.2,
-        promotion_eligibility: true,
-        color_code: "#FFD700",
-        sort_order: 2,
-        is_active: true,
-        code: null,
-        name: null,
-      },
-      {
-        grade_code: "ADVANCED",
-        grade_name: "上級",
-        grade_name_short: "上級",
-        grade_level: 3,
-        description: "複雑な業務をリードし、チーム全体の技術指導ができるレベル",
-        evaluation_criteria: "高度な技術力、リーダーシップ、戦略的思考",
-        required_experience_months: 36,
-        skill_indicators: "["高度技術", "リーダーシップ", "戦略思考"]",
-        competency_requirements: "["専門技術", "チーム管理", "技術戦略立案"]",
-        certification_requirements: "["応用情報技術者", "専門資格"]",
-        project_complexity: "COMPLEX",
-        mentoring_capability: true,
-        leadership_level: "PROJECT",
-        salary_impact_factor: 1.5,
-        promotion_eligibility: true,
-        color_code: "#FF8C00",
+    }),
+    prisma.position.upsert({
+      where: { position_code: 'POS003' },
+      update: {},
+      create: {
+        position_code: 'POS003',
+        position_name: '主任',
+        position_level: 2,
         sort_order: 3,
-        is_active: true,
-        code: null,
-        name: null,
+        code: 'POS003',
+        name: '主任',
       },
-      {
-        grade_code: "EXPERT",
-        grade_name: "専門家",
-        grade_name_short: "専門",
-        grade_level: 4,
-        description: "組織全体の技術方針に影響を与え、業界レベルでの専門性を持つレベル",
-        evaluation_criteria: "業界専門性、組織への影響力、イノベーション創出",
-        required_experience_months: 60,
-        skill_indicators: "["業界専門性", "組織影響力", "イノベーション"]",
-        competency_requirements: "["業界知識", "組織運営", "技術革新"]",
-        certification_requirements: "["高度情報技術者", "業界認定資格"]",
-        project_complexity: "CRITICAL",
-        mentoring_capability: true,
-        leadership_level: "ORGANIZATION",
-        salary_impact_factor: 2.0,
-        promotion_eligibility: true,
-        color_code: "#DC143C",
+    }),
+    prisma.position.upsert({
+      where: { position_code: 'POS004' },
+      update: {},
+      create: {
+        position_code: 'POS004',
+        position_name: '一般',
+        position_level: 1,
         sort_order: 4,
-        is_active: true,
-        code: null,
-        name: null,
+        code: 'POS004',
+        name: '一般',
       },
-    ],
-  })
-  
-  // MST_SkillHierarchy_sample_data.sql
-  console.log('📊 MST_SkillHierarchyデータを投入中...')
-  await prisma.skillHierarchy.createMany({
-    data: [
-      {
-        skill_id: "SKILL_TECH_001",
-        parent_skill_id: null,
-        hierarchy_level: 1,
-        skill_path: "/技術スキル",
-        sort_order: 1,
-        is_leaf: false,
-        skill_category: "TECHNICAL",
-        description: "技術系スキルの大分類",
-        is_active: true,
-        code: null,
-        name: null,
+    }),
+  ])
+
+  // 4. 権限マスターデータ
+  console.log('🔐 権限マスターデータを投入中...')
+  const roles = await Promise.all([
+    prisma.role.upsert({
+      where: { role_code: 'ROLE001' },
+      update: {},
+      create: {
+        role_code: 'ROLE001',
+        role_name: 'システム管理者',
+        role_category: 'ADMIN',
+        role_level: 5,
+        is_system_role: true,
+        role_status: 'ACTIVE',
+        code: 'ROLE001',
+        name: 'システム管理者',
+        description: 'システム全体の管理権限',
       },
-      {
-        skill_id: "SKILL_PROG_001",
-        parent_skill_id: "SKILL_TECH_001",
-        hierarchy_level: 2,
-        skill_path: "/技術スキル/プログラミング",
-        sort_order: 1,
-        is_leaf: false,
-        skill_category: "TECHNICAL",
-        description: "プログラミング言語・技術",
-        is_active: true,
-        code: null,
-        name: null,
+    }),
+    prisma.role.upsert({
+      where: { role_code: 'ROLE002' },
+      update: {},
+      create: {
+        role_code: 'ROLE002',
+        role_name: '部門管理者',
+        role_category: 'MANAGER',
+        role_level: 3,
+        is_system_role: false,
+        role_status: 'ACTIVE',
+        code: 'ROLE002',
+        name: '部門管理者',
+        description: '部門内のデータ管理権限',
       },
-      {
-        skill_id: "SKILL_JAVA_001",
-        parent_skill_id: "SKILL_PROG_001",
-        hierarchy_level: 3,
-        skill_path: "/技術スキル/プログラミング/Java",
-        sort_order: 1,
-        is_leaf: true,
-        skill_category: "TECHNICAL",
-        description: "Java言語でのプログラミングスキル",
-        is_active: true,
-        code: null,
-        name: null,
+    }),
+    prisma.role.upsert({
+      where: { role_code: 'ROLE003' },
+      update: {},
+      create: {
+        role_code: 'ROLE003',
+        role_name: '一般ユーザー',
+        role_category: 'USER',
+        role_level: 1,
+        is_system_role: false,
+        role_status: 'ACTIVE',
+        code: 'ROLE003',
+        name: '一般ユーザー',
+        description: '自分のデータのみ編集可能',
       },
-    ],
-  })
-  
-  // MST_SkillItem_sample_data.sql
-  console.log('📊 MST_SkillItemデータを投入中...')
-  await prisma.skillItem.createMany({
-    data: [
-      {
-        skill_code: "SKILL001",
-        skill_name: "Java",
-        skill_category_id: "CAT001",
-        skill_type: "TECHNICAL",
+    }),
+  ])
+
+  // 5. 権限設定データ
+  console.log('🔑 権限設定データを投入中...')
+  const permissions = await Promise.all([
+    prisma.permission.upsert({
+      where: { permission_code: 'PERM001' },
+      update: {},
+      create: {
+        permission_code: 'PERM001',
+        permission_name: 'ユーザー情報閲覧',
+        permission_category: 'USER',
+        resource_type: 'USER',
+        action_type: 'READ',
+        scope_level: 'ALL',
+        permission_status: 'ACTIVE',
+        code: 'PERM001',
+        name: 'ユーザー情報閲覧',
+        description: 'ユーザー情報の閲覧権限',
+      },
+    }),
+    prisma.permission.upsert({
+      where: { permission_code: 'PERM002' },
+      update: {},
+      create: {
+        permission_code: 'PERM002',
+        permission_name: 'ユーザー情報編集',
+        permission_category: 'USER',
+        resource_type: 'USER',
+        action_type: 'WRITE',
+        scope_level: 'SELF',
+        permission_status: 'ACTIVE',
+        code: 'PERM002',
+        name: 'ユーザー情報編集',
+        description: 'ユーザー情報の編集権限',
+      },
+    }),
+    prisma.permission.upsert({
+      where: { permission_code: 'PERM003' },
+      update: {},
+      create: {
+        permission_code: 'PERM003',
+        permission_name: 'スキル情報閲覧',
+        permission_category: 'SKILL',
+        resource_type: 'SKILL',
+        action_type: 'READ',
+        scope_level: 'ALL',
+        permission_status: 'ACTIVE',
+        code: 'PERM003',
+        name: 'スキル情報閲覧',
+        description: 'スキル情報の閲覧権限',
+      },
+    }),
+    prisma.permission.upsert({
+      where: { permission_code: 'PERM004' },
+      update: {},
+      create: {
+        permission_code: 'PERM004',
+        permission_name: 'スキル情報編集',
+        permission_category: 'SKILL',
+        resource_type: 'SKILL',
+        action_type: 'WRITE',
+        scope_level: 'SELF',
+        permission_status: 'ACTIVE',
+        code: 'PERM004',
+        name: 'スキル情報編集',
+        description: 'スキル情報の編集権限',
+      },
+    }),
+    prisma.permission.upsert({
+      where: { permission_code: 'PERM005' },
+      update: {},
+      create: {
+        permission_code: 'PERM005',
+        permission_name: '管理者権限',
+        permission_category: 'ADMIN',
+        resource_type: 'SYSTEM',
+        action_type: 'ALL',
+        scope_level: 'ALL',
+        permission_status: 'ACTIVE',
+        code: 'PERM005',
+        name: '管理者権限',
+        description: '管理者権限（全操作）',
+      },
+    }),
+  ])
+
+  // 6. スキルカテゴリマスターデータ
+  console.log('🎯 スキルカテゴリマスターデータを投入中...')
+  const skillCategories = await Promise.all([
+    prisma.skillCategory.upsert({
+      where: { category_code: 'CAT001' },
+      update: {},
+      create: {
+        category_code: 'CAT001',
+        category_name: 'フロントエンド',
+        category_type: 'TECHNICAL',
+        parent_category_id: null,
+        category_level: 1,
+        display_order: 1,
+        category_status: 'ACTIVE',
+        code: 'CAT001',
+        name: 'フロントエンド',
+      },
+    }),
+    prisma.skillCategory.upsert({
+      where: { category_code: 'CAT002' },
+      update: {},
+      create: {
+        category_code: 'CAT002',
+        category_name: 'バックエンド',
+        category_type: 'TECHNICAL',
+        parent_category_id: null,
+        category_level: 1,
+        display_order: 2,
+        category_status: 'ACTIVE',
+        code: 'CAT002',
+        name: 'バックエンド',
+      },
+    }),
+    prisma.skillCategory.upsert({
+      where: { category_code: 'CAT003' },
+      update: {},
+      create: {
+        category_code: 'CAT003',
+        category_name: 'インフラ',
+        category_type: 'TECHNICAL',
+        parent_category_id: null,
+        category_level: 1,
+        display_order: 3,
+        category_status: 'ACTIVE',
+        code: 'CAT003',
+        name: 'インフラ',
+      },
+    }),
+    prisma.skillCategory.upsert({
+      where: { category_code: 'CAT004' },
+      update: {},
+      create: {
+        category_code: 'CAT004',
+        category_name: 'マネジメント',
+        category_type: 'BUSINESS',
+        parent_category_id: null,
+        category_level: 1,
+        display_order: 4,
+        category_status: 'ACTIVE',
+        code: 'CAT004',
+        name: 'マネジメント',
+      },
+    }),
+  ])
+
+  // 7. スキルマスターデータ
+  console.log('⚡ スキルマスターデータを投入中...')
+  const skills = await Promise.all([
+    // フロントエンドスキル
+    prisma.skillItem.upsert({
+      where: { skill_code: 'SKL001' },
+      update: {},
+      create: {
+        skill_code: 'SKL001',
+        skill_name: 'React.js',
+        skill_category_id: skillCategories[0].category_code,
+        skill_type: 'FRAMEWORK',
+        difficulty_level: 3,
+        importance_level: 5,
+        code: 'SKL001',
+        name: 'React.js',
+        description: 'Reactライブラリを使用したフロントエンド開発',
+      },
+    }),
+    prisma.skillItem.upsert({
+      where: { skill_code: 'SKL002' },
+      update: {},
+      create: {
+        skill_code: 'SKL002',
+        skill_name: 'Next.js',
+        skill_category_id: skillCategories[0].category_code,
+        skill_type: 'FRAMEWORK',
+        difficulty_level: 4,
+        importance_level: 5,
+        code: 'SKL002',
+        name: 'Next.js',
+        description: 'Next.jsフレームワークを使用したフルスタック開発',
+      },
+    }),
+    prisma.skillItem.upsert({
+      where: { skill_code: 'SKL003' },
+      update: {},
+      create: {
+        skill_code: 'SKL003',
+        skill_name: 'TypeScript',
+        skill_category_id: skillCategories[0].category_code,
+        skill_type: 'LANGUAGE',
         difficulty_level: 3,
         importance_level: 4,
-        code: null,
-        name: null,
-        description: null,
+        code: 'SKL003',
+        name: 'TypeScript',
+        description: 'TypeScriptを使用した型安全な開発',
       },
-    ],
-  })
-  
-  // MST_SystemConfig_sample_data.sql
-  console.log('📊 MST_SystemConfigデータを投入中...')
-  await prisma.systemConfig.createMany({
-    data: [
-      {
-        config_key: "MAX_LOGIN_ATTEMPTS",
-        config_name: "最大ログイン試行回数",
-        config_value: 5,
-        config_type: "INTEGER",
-        config_category: "SECURITY",
-        default_value: 3,
-        validation_rule: "^[1-9][0-9]*$",
-        description: "アカウントロックまでの最大ログイン失敗回数",
-        is_encrypted: false,
-        is_system_only: false,
-        is_user_configurable: true,
-        requires_restart: false,
-        environment: "ALL",
-        tenant_specific: true,
-        last_modified_by: "admin",
-        last_modified_reason: "セキュリティ強化のため",
-        sort_order: 1,
-        is_active: true,
-        code: null,
-        name: null,
+    }),
+    // バックエンドスキル
+    prisma.skillItem.upsert({
+      where: { skill_code: 'SKL004' },
+      update: {},
+      create: {
+        skill_code: 'SKL004',
+        skill_name: 'Node.js',
+        skill_category_id: skillCategories[1].category_code,
+        skill_type: 'RUNTIME',
+        difficulty_level: 3,
+        importance_level: 4,
+        code: 'SKL004',
+        name: 'Node.js',
+        description: 'Node.jsを使用したサーバーサイド開発',
       },
-      {
-        config_key: "SESSION_TIMEOUT_MINUTES",
-        config_name: "セッションタイムアウト時間（分）",
-        config_value: 30,
-        config_type: "INTEGER",
-        config_category: "SECURITY",
-        default_value: 60,
-        validation_rule: "^[1-9][0-9]*$",
-        description: "ユーザーセッションの自動タイムアウト時間",
-        is_encrypted: false,
-        is_system_only: false,
-        is_user_configurable: true,
-        requires_restart: false,
-        environment: "ALL",
-        tenant_specific: true,
-        last_modified_by: "admin",
-        last_modified_reason: "セキュリティポリシー変更",
-        sort_order: 2,
-        is_active: true,
-        code: null,
-        name: null,
+    }),
+    prisma.skillItem.upsert({
+      where: { skill_code: 'SKL005' },
+      update: {},
+      create: {
+        skill_code: 'SKL005',
+        skill_name: 'PostgreSQL',
+        skill_category_id: skillCategories[1].category_code,
+        skill_type: 'DATABASE',
+        difficulty_level: 3,
+        importance_level: 4,
+        code: 'SKL005',
+        name: 'PostgreSQL',
+        description: 'PostgreSQLデータベースの設計・運用',
       },
-      {
-        config_key: "SKILL_EVALUATION_PERIOD_MONTHS",
-        config_name: "スキル評価期間（月）",
-        config_value: 6,
-        config_type: "INTEGER",
-        config_category: "BUSINESS",
-        default_value: 12,
-        validation_rule: "^[1-9][0-9]*$",
-        description: "スキル評価の実施間隔",
-        is_encrypted: false,
-        is_system_only: false,
-        is_user_configurable: true,
-        requires_restart: false,
-        environment: "ALL",
-        tenant_specific: true,
-        last_modified_by: "hr_admin",
-        last_modified_reason: "評価頻度の見直し",
-        sort_order: 10,
-        is_active: true,
-        code: null,
-        name: null,
+    }),
+    // インフラスキル
+    prisma.skillItem.upsert({
+      where: { skill_code: 'SKL006' },
+      update: {},
+      create: {
+        skill_code: 'SKL006',
+        skill_name: 'Docker',
+        skill_category_id: skillCategories[2].category_code,
+        skill_type: 'TOOL',
+        difficulty_level: 3,
+        importance_level: 3,
+        code: 'SKL006',
+        name: 'Docker',
+        description: 'Dockerを使用したコンテナ化技術',
       },
-      {
-        config_key: "EMAIL_SMTP_PASSWORD",
-        config_name: "SMTP認証パスワード",
-        config_value: "encrypted_password_value",
-        config_type: "ENCRYPTED",
-        config_category: "INTEGRATION",
-        default_value: null,
-        validation_rule: null,
-        description: "メール送信用SMTP認証パスワード",
-        is_encrypted: true,
-        is_system_only: true,
-        is_user_configurable: false,
-        requires_restart: true,
-        environment: "PROD",
-        tenant_specific: false,
-        last_modified_by: "system",
-        last_modified_reason: "初期設定",
-        sort_order: 100,
-        is_active: true,
-        code: null,
-        name: null,
+    }),
+    prisma.skillItem.upsert({
+      where: { skill_code: 'SKL007' },
+      update: {},
+      create: {
+        skill_code: 'SKL007',
+        skill_name: 'Vercel',
+        skill_category_id: skillCategories[2].category_code,
+        skill_type: 'PLATFORM',
+        difficulty_level: 2,
+        importance_level: 3,
+        code: 'SKL007',
+        name: 'Vercel',
+        description: 'Vercelを使用したデプロイメント',
       },
-    ],
-  })
-  
-  // MST_TenantSettings_sample_data.sql
-  console.log('📊 MST_TenantSettingsデータを投入中...')
-  await prisma.tenantSettings.createMany({
-    data: [
-      {
-        id: "TS001",
-        tenant_id: "TENANT001",
-        setting_category: "SYSTEM",
-        setting_key: "max_users",
-        setting_name: "最大ユーザー数",
-        setting_description: "このテナントで作成可能な最大ユーザー数",
-        data_type: "INTEGER",
-        setting_value: 100,
-        default_value: 50,
-        validation_rules: "{"min": 1, "max": 1000}",
-        is_required: true,
-        is_encrypted: false,
-        is_system_managed: false,
-        is_user_configurable: false,
-        display_order: 1,
-        effective_from: "2025-01-01 00:00:00",
-        effective_until: null,
-        last_modified_by: "SYSTEM",
-        code: null,
-        name: null,
-        description: null,
+    }),
+  ])
+
+  // 8. 従業員マスターデータ
+  console.log('👤 従業員マスターデータを投入中...')
+  const employees = await Promise.all([
+    // 管理者ユーザー
+    prisma.employee.upsert({
+      where: { employee_code: 'ADMIN001' },
+      update: {},
+      create: {
+        employee_code: 'ADMIN001',
+        full_name: 'システム管理者',
+        full_name_kana: 'システムカンリシャ',
+        email: 'admin@skill-report.local',
+        hire_date: new Date('2025-01-01'),
+        department_id: departments[0].department_code,
+        position_id: positions[0].position_code,
+        employment_status: 'ACTIVE',
+        employee_status: 'ACTIVE',
+        code: 'ADMIN001',
+        name: 'システム管理者',
       },
-      {
-        id: "TS002",
-        tenant_id: "TENANT001",
-        setting_category: "UI",
-        setting_key: "theme_color",
-        setting_name: "テーマカラー",
-        setting_description: "システムのメインテーマカラー",
-        data_type: "STRING",
-        setting_value: "#2563eb",
-        default_value: "#3b82f6",
-        validation_rules: "{"pattern": "^#[0-9a-fA-F]{6}$"}",
-        is_required: false,
-        is_encrypted: false,
-        is_system_managed: false,
-        is_user_configurable: true,
-        display_order: 1,
-        effective_from: null,
-        effective_until: null,
-        last_modified_by: "USER001",
-        code: null,
-        name: null,
-        description: null,
+    }),
+    // テストユーザー
+    prisma.employee.upsert({
+      where: { employee_code: 'TEST001' },
+      update: {},
+      create: {
+        employee_code: 'TEST001',
+        full_name: 'テスト太郎',
+        full_name_kana: 'テストタロウ',
+        email: 'test@skill-report.local',
+        hire_date: new Date('2025-04-01'),
+        department_id: departments[0].department_code,
+        position_id: positions[3].position_code,
+        employment_status: 'ACTIVE',
+        employee_status: 'ACTIVE',
+        code: 'TEST001',
+        name: 'テスト太郎',
       },
-      {
-        id: "TS003",
-        tenant_id: "TENANT001",
-        setting_category: "BUSINESS",
-        setting_key: "skill_approval_required",
-        setting_name: "スキル承認必須",
-        setting_description: "スキル登録時に承認が必要かどうか",
-        data_type: "BOOLEAN",
-        setting_value: "true",
-        default_value: "false",
-        validation_rules: null,
-        is_required: true,
-        is_encrypted: false,
-        is_system_managed: false,
-        is_user_configurable: true,
-        display_order: 1,
-        effective_from: null,
-        effective_until: null,
-        last_modified_by: "USER001",
-        code: null,
-        name: null,
-        description: null,
-      },
-      {
-        id: "TS004",
-        tenant_id: "TENANT001",
-        setting_category: "SECURITY",
-        setting_key: "password_policy",
-        setting_name: "パスワードポリシー",
-        setting_description: "パスワードの複雑性要件",
-        data_type: "JSON",
-        setting_value: "{"min_length": 8, "require_uppercase": true, "require_lowercase": true, "require_numbers": true, "require_symbols": false}",
-        default_value: "{"min_length": 6, "require_uppercase": false, "require_lowercase": false, "require_numbers": false, "require_symbols": false}",
-        validation_rules: "{"type": "object", "properties": {"min_length": {"type": "integer", "minimum": 4, "maximum": 128}}}",
-        is_required: true,
-        is_encrypted: false,
-        is_system_managed: false,
-        is_user_configurable: true,
-        display_order: 1,
-        effective_from: null,
-        effective_until: null,
-        last_modified_by: "USER001",
-        code: null,
-        name: null,
-        description: null,
-      },
-    ],
-  })
-  
-  // MST_Tenant_sample_data.sql
-  console.log('📊 MST_Tenantデータを投入中...')
-  await prisma.tenant.createMany({
-    data: [
-      {
-        tenant_id: "TENANT_001",
-        tenant_code: "acme-corp",
-        tenant_name: "株式会社ACME",
-        tenant_name_en: "ACME Corporation",
-        tenant_short_name: "ACME",
-        tenant_type: "ENTERPRISE",
-        parent_tenant_id: null,
-        tenant_level: 1,
-        domain_name: "acme-corp.com",
-        subdomain: "acme",
-        logo_url: "https://cdn.example.com/logos/acme-corp.png",
-        primary_color: "#0066CC",
-        secondary_color: "#FF6600",
-        timezone: "Asia/Tokyo",
-        locale: "ja_JP",
-        currency_code: "JPY",
-        date_format: "YYYY-MM-DD",
-        time_format: "HH:mm:ss",
-        admin_email: "admin@acme-corp.com",
-        contact_email: "contact@acme-corp.com",
-        phone_number: "03-1234-5678",
-        address: "東京都千代田区丸の内1-1-1",
-        postal_code: "100-0005",
-        country_code: "JP",
-        subscription_plan: "ENTERPRISE",
-        max_users: 1000,
-        max_storage_gb: 1000,
-        features_enabled: "["advanced_analytics", "custom_reports", "api_access", "sso", "audit_logs"]",
-        custom_settings: "{"theme": "corporate", "dashboard_layout": "advanced", "notification_preferences": {"email": true, "slack": true}}",
-        security_policy: "{"password_policy": {"min_length": 8, "require_special_chars": true}, "session_timeout": 480, "ip_whitelist": ["192.168.1.0/24"]}",
-        data_retention_days: 2555,
-        backup_enabled: true,
-        backup_frequency: "DAILY",
-        contract_start_date: "2024-01-01",
-        contract_end_date: "2024-12-31",
-        trial_end_date: null,
-        billing_cycle: "ANNUAL",
-        monthly_fee: 50000.0,
-        setup_fee: 100000.0,
-        status: "ACTIVE",
-        activation_date: "2024-01-01",
-        suspension_date: null,
-        suspension_reason: null,
-        last_login_date: "2024-06-01",
-        current_users_count: 250,
-        storage_used_gb: 125.5,
-        api_rate_limit: 10000,
-        sso_enabled: true,
-        sso_provider: "SAML",
-        sso_config: "{"entity_id": "acme-corp", "sso_url": "https://sso.acme-corp.com/saml", "certificate": "..."}",
-        webhook_url: "https://api.acme-corp.com/webhooks/skill-system",
-        webhook_secret: "webhook_secret_key_123",
-        created_by: "SYSTEM",
-        notes: "大手企業向けエンタープライズプラン",
-        code: null,
-        name: null,
-        description: null,
-      },
-      {
-        tenant_id: "TENANT_002",
-        tenant_code: "beta-tech",
-        tenant_name: "ベータテクノロジー株式会社",
-        tenant_name_en: "Beta Technology Inc.",
-        tenant_short_name: "BetaTech",
-        tenant_type: "ENTERPRISE",
-        parent_tenant_id: null,
-        tenant_level: 1,
-        domain_name: null,
-        subdomain: "beta-tech",
-        logo_url: "https://cdn.example.com/logos/beta-tech.png",
-        primary_color: "#28A745",
-        secondary_color: "#6C757D",
-        timezone: "Asia/Tokyo",
-        locale: "ja_JP",
-        currency_code: "JPY",
-        date_format: "YYYY/MM/DD",
-        time_format: "HH:mm",
-        admin_email: "admin@beta-tech.co.jp",
-        contact_email: "info@beta-tech.co.jp",
-        phone_number: "06-9876-5432",
-        address: "大阪府大阪市北区梅田2-2-2",
-        postal_code: "530-0001",
-        country_code: "JP",
-        subscription_plan: "STANDARD",
-        max_users: 200,
-        max_storage_gb: 100,
-        features_enabled: "["basic_analytics", "standard_reports", "api_access"]",
-        custom_settings: "{"theme": "modern", "dashboard_layout": "standard"}",
-        security_policy: "{"password_policy": {"min_length": 6, "require_special_chars": false}, "session_timeout": 240}",
-        data_retention_days: 1825,
-        backup_enabled: true,
-        backup_frequency: "WEEKLY",
-        contract_start_date: "2024-03-01",
-        contract_end_date: "2025-02-28",
-        trial_end_date: null,
-        billing_cycle: "MONTHLY",
-        monthly_fee: 15000.0,
-        setup_fee: 30000.0,
-        status: "ACTIVE",
-        activation_date: "2024-03-01",
-        suspension_date: null,
-        suspension_reason: null,
-        last_login_date: "2024-05-30",
-        current_users_count: 85,
-        storage_used_gb: 23.75,
-        api_rate_limit: 2000,
-        sso_enabled: false,
-        sso_provider: null,
-        sso_config: null,
-        webhook_url: null,
-        webhook_secret: null,
-        created_by: "SYSTEM",
-        notes: "中堅企業向けスタンダードプラン",
-        code: null,
-        name: null,
-        description: null,
-      },
-    ],
-  })
-  
-  // MST_TrainingProgram_sample_data.sql
-  console.log('📊 MST_TrainingProgramデータを投入中...')
-  await prisma.trainingProgram.createMany({
-    data: [
-      {
-        training_program_id: "TRN_PROG_001",
-        program_code: "PM-BASIC-001",
-        program_name: "プロジェクトマネジメント基礎研修",
-        program_name_en: "Project Management Fundamentals",
-        program_description: "プロジェクトマネジメントの基本概念と手法を学ぶ研修",
-        program_category: "MANAGEMENT",
-        program_type: "CLASSROOM",
-        target_audience: "MIDDLE",
-        difficulty_level: "INTERMEDIATE",
-        duration_hours: 16.0,
-        duration_days: 2,
-        max_participants: 20,
-        min_participants: 8,
-        prerequisites: "実務経験2年以上、基本的なビジネススキル",
-        learning_objectives: "PMBOKの基本概念理解、プロジェクト計画立案、リスク管理手法の習得",
-        curriculum_outline: "1日目：PM概論、計画立案　2日目：実行・監視、リスク管理",
-        curriculum_details: "{"day1": ["PM概論", "プロジェクト憲章", "WBS作成"], "day2": ["進捗管理", "リスク分析", "ケーススタディ"]}",
-        materials_required: "["テキスト", "演習用PC", "プロジェクト計画テンプレート"]",
-        equipment_required: "["プロジェクター", "ホワイトボード", "PC環境"]",
-        instructor_requirements: "PMP資格保有、実務経験5年以上",
-        assessment_method: "COMPREHENSIVE",
-        passing_score: 70.0,
-        certification_provided: true,
-        pdu_credits: 16.0,
-        related_skills: "["プロジェクト管理", "リーダーシップ", "コミュニケーション"]",
-        related_certifications: "["PMP", "プロジェクトマネージャ試験"]",
-        cost_per_participant: 50000.0,
-        external_provider: null,
-        external_url: null,
-        venue_type: "INTERNAL",
-        venue_requirements: "20名収容可能な研修室、プロジェクター設備",
-        language: "JA",
-        repeat_interval: 24,
-        mandatory_flag: false,
-        active_flag: true,
-        effective_start_date: "2024-01-01",
-        effective_end_date: null,
-        created_by: "EMP000010",
-        approved_by: "EMP000005",
-        approval_date: "2023-12-15",
-        version_number: 1.0,
-        revision_notes: "初版作成",
-        tags: "["プロジェクト管理", "PMBOK", "リーダーシップ", "中級"]",
-        code: null,
-        name: null,
-        description: null,
-      },
-      {
-        training_program_id: "TRN_PROG_002",
-        program_code: "AWS-ARCH-001",
-        program_name: "AWS認定ソリューションアーキテクト対策研修",
-        program_name_en: "AWS Certified Solutions Architect Preparation",
-        program_description: "AWS認定ソリューションアーキテクト資格取得のための対策研修",
-        program_category: "TECHNICAL",
-        program_type: "BLENDED",
-        target_audience: "SENIOR",
-        difficulty_level: "ADVANCED",
-        duration_hours: 24.0,
-        duration_days: 3,
-        max_participants: 15,
-        min_participants: 5,
-        prerequisites: "AWS基礎知識、クラウド実務経験1年以上",
-        learning_objectives: "AWSサービス理解、アーキテクチャ設計、試験合格",
-        curriculum_outline: "1日目：AWS基礎　2日目：アーキテクチャ設計　3日目：模擬試験・解説",
-        curriculum_details: "{"day1": ["EC2", "S3", "VPC"], "day2": ["高可用性設計", "セキュリティ", "コスト最適化"], "day3": ["模擬試験", "解説", "試験対策"]}",
-        materials_required: "["AWS公式テキスト", "模擬試験問題集", "ハンズオン環境"]",
-        equipment_required: "["AWS環境", "PC", "インターネット接続"]",
-        instructor_requirements: "AWS認定資格保有、実務経験3年以上",
-        assessment_method: "TEST",
-        passing_score: 80.0,
-        certification_provided: true,
-        pdu_credits: 24.0,
-        related_skills: "["AWS", "クラウドアーキテクチャ", "インフラ設計"]",
-        related_certifications: "["AWS認定ソリューションアーキテクト"]",
-        cost_per_participant: 80000.0,
-        external_provider: "AWSトレーニングパートナー",
-        external_url: "https://aws.amazon.com/training/",
-        venue_type: "HYBRID",
-        venue_requirements: "PC環境、AWS環境アクセス可能",
-        language: "JA",
-        repeat_interval: 12,
-        mandatory_flag: false,
-        active_flag: true,
-        effective_start_date: "2024-02-01",
-        effective_end_date: null,
-        created_by: "EMP000015",
-        approved_by: "EMP000008",
-        approval_date: "2024-01-20",
-        version_number: 1.1,
-        revision_notes: "ハンズオン内容を強化",
-        tags: "["AWS", "クラウド", "認定資格", "アーキテクチャ", "上級"]",
-        code: null,
-        name: null,
-        description: null,
-      },
-    ],
-  })
-  
-  // MST_UserAuth_sample_data.sql
-  console.log('📊 MST_UserAuthデータを投入中...')
-  await prisma.userAuth.createMany({
-    data: [
-      {
-        user_id: "USER000001",
-        login_id: "yamada.taro@company.com",
-        password_hash: "$2b$12$LQv3c1yqBWVHxkd0LHAkCOYz6TtxMQJqhN8/LewdBPj/RK.s5uIoS",
-        password_salt: "randomsalt123",
-        employee_id: "EMP000001",
-        account_status: "ACTIVE",
-        last_login_at: "2025-06-01 09:00:00",
-        last_login_ip: "192.168.1.100",
+    }),
+  ])
+
+  // 9. 認証情報の作成
+  console.log('🔐 認証情報を作成中...')
+  const userAuths = await Promise.all([
+    // 管理者の認証情報
+    prisma.userAuth.upsert({
+      where: { user_id: employees[0].employee_code },
+      update: {},
+      create: {
+        user_id: employees[0].employee_code,
+        login_id: 'admin@skill-report.local',
+        password_hash: '$2a$10$92IXUNpkjO0rOQ5byMi.Ye4oKoEa3Ro9llC/.og/at2.uheWG/igi', // password
+        employee_id: employees[0].employee_code,
+        account_status: 'ACTIVE',
         failed_login_count: 0,
-        last_failed_login_at: null,
-        password_changed_at: "2025-01-01 00:00:00",
-        password_expires_at: "2025-12-31 23:59:59",
-        mfa_enabled: true,
-        mfa_secret: "JBSWY3DPEHPK3PXP",
-        recovery_token: null,
-        recovery_token_expires_at: null,
-        session_timeout: 480,
-        external_auth_provider: null,
-        external_auth_id: null,
-        code: null,
-        name: null,
-        description: null,
-      },
-      {
-        user_id: "USER000002",
-        login_id: "sato.hanako@company.com",
-        password_hash: "$2b$12$XQv3c1yqBWVHxkd0LHAkCOYz6TtxMQJqhN8/LewdBPj/RK.s5uIoX",
-        password_salt: "randomsalt456",
-        employee_id: "EMP000002",
-        account_status: "ACTIVE",
-        last_login_at: "2025-05-31 17:30:00",
-        last_login_ip: "192.168.1.101",
-        failed_login_count: 0,
-        last_failed_login_at: null,
-        password_changed_at: "2025-02-01 00:00:00",
-        password_expires_at: "2026-01-31 23:59:59",
         mfa_enabled: false,
-        mfa_secret: null,
-        recovery_token: null,
-        recovery_token_expires_at: null,
-        session_timeout: 240,
-        external_auth_provider: null,
-        external_auth_id: null,
-        code: null,
-        name: null,
-        description: null,
+        code: 'AUTH001',
+        name: 'システム管理者認証',
       },
-    ],
-  })
-  
-  // MST_UserRole_sample_data.sql
-  console.log('📊 MST_UserRoleデータを投入中...')
-  await prisma.userRole.createMany({
-    data: [
-      {
-        user_id: "USER000001",
-        role_id: "ROLE003",
-        assignment_type: "DIRECT",
-        assigned_by: "USER000000",
-        assignment_reason: "新規ユーザー登録時の標準ロール割り当て",
-        effective_from: "2025-01-01 00:00:00",
-        effective_to: null,
-        is_primary_role: true,
-        priority_order: 1,
-        conditions: null,
-        delegation_source_user_id: null,
-        delegation_expires_at: null,
-        auto_assigned: true,
-        requires_approval: false,
-        approval_status: null,
-        approved_by: null,
-        approved_at: null,
-        assignment_status: "ACTIVE",
-        last_used_at: "2025-06-01 09:00:00",
-        usage_count: 150,
-        code: null,
-        name: null,
-        description: null,
+    }),
+    // テストユーザーの認証情報
+    prisma.userAuth.upsert({
+      where: { user_id: employees[1].employee_code },
+      update: {},
+      create: {
+        user_id: employees[1].employee_code,
+        login_id: 'test@skill-report.local',
+        password_hash: '$2a$10$92IXUNpkjO0rOQ5byMi.Ye4oKoEa3Ro9llC/.og/at2.uheWG/igi', // password
+        employee_id: employees[1].employee_code,
+        account_status: 'ACTIVE',
+        failed_login_count: 0,
+        mfa_enabled: false,
+        code: 'AUTH002',
+        name: 'テストユーザー認証',
       },
-      {
-        user_id: "USER000002",
-        role_id: "ROLE002",
-        assignment_type: "DIRECT",
-        assigned_by: "USER000001",
-        assignment_reason: "テナント管理者権限付与",
-        effective_from: "2025-02-01 00:00:00",
-        effective_to: null,
-        is_primary_role: true,
-        priority_order: 1,
-        conditions: "{"tenant_id": "TENANT001"}",
-        delegation_source_user_id: null,
-        delegation_expires_at: null,
-        auto_assigned: false,
-        requires_approval: true,
-        approval_status: "APPROVED",
-        approved_by: "USER000001",
-        approved_at: "2025-01-31 15:30:00",
-        assignment_status: "ACTIVE",
-        last_used_at: "2025-06-01 10:30:00",
-        usage_count: 75,
-        code: null,
-        name: null,
-        description: null,
-      },
-    ],
-  })
-  
-  // SYS_BackupHistory_sample_data.sql
-  console.log('📊 SYS_BackupHistoryデータを投入中...')
-  await prisma.backupHistory.createMany({
-    data: [
-      {
-        backup_id: "BKP_20240101_001",
-        backup_type: "FULL",
-        backup_scope: "DATABASE",
-        target_objects: null,
-        backup_start_time: "2024-01-01 02:00:00",
-        backup_end_time: "2024-01-01 02:45:00",
-        backup_status: "SUCCESS",
-        backup_file_path: "/backup/full/skill_report_20240101_020000.sql.gz",
-        backup_file_size: 1073741824,
-        compression_type: "GZIP",
-        encryption_enabled: true,
-        checksum: "a1b2c3d4e5f6789012345678901234567890abcd",
-        retention_period_days: 90,
-        expiry_date: "2024-04-01",
-        backup_trigger: "SCHEDULED",
-        executed_by: "system_backup_job",
-        error_message: null,
-        recovery_tested: true,
-        recovery_test_date: "2024-01-15",
-        id: null,
-        is_deleted: null,
-      },
-      {
-        backup_id: "BKP_20240101_002",
-        backup_type: "INCREMENTAL",
-        backup_scope: "DATABASE",
-        target_objects: null,
-        backup_start_time: "2024-01-01 14:00:00",
-        backup_end_time: "2024-01-01 14:05:00",
-        backup_status: "SUCCESS",
-        backup_file_path: "/backup/incremental/skill_report_20240101_140000.sql.gz",
-        backup_file_size: 52428800,
-        compression_type: "GZIP",
-        encryption_enabled: true,
-        checksum: "b2c3d4e5f6789012345678901234567890abcde1",
-        retention_period_days: 30,
-        expiry_date: "2024-01-31",
-        backup_trigger: "SCHEDULED",
-        executed_by: "system_backup_job",
-        error_message: null,
-        recovery_tested: false,
-        recovery_test_date: null,
-        id: null,
-        is_deleted: null,
-      },
-    ],
-  })
-  
-  // SYS_IntegrationConfig_sample_data.sql
-  console.log('📊 SYS_IntegrationConfigデータを投入中...')
-  await prisma.integrationConfig.createMany({
-    data: [
-      {
-        id: "IC001",
-        tenant_id: "TENANT001",
-        integration_key: "slack_webhook",
-        integration_name: "Slack通知連携",
-        integration_type: "WEBHOOK",
-        endpoint_url: "https://hooks.slack.com/services/T00000000/B00000000/XXXXXXXXXXXXXXXXXXXXXXXX",
-        auth_type: "NONE",
-        auth_config: null,
-        connection_config: "{"channel": "#notifications", "username": "SkillBot", "icon_emoji": ":robot_face:"}",
-        request_headers: "{"Content-Type": "application/json"}",
-        timeout_seconds: 30,
-        retry_count: 3,
-        retry_interval: 5,
-        rate_limit_per_minute: 60,
-        is_enabled: true,
-        health_check_url: null,
-        last_health_check: null,
-        health_status: "UNKNOWN",
-        is_deleted: null,
-      },
-      {
-        id: "IC002",
-        tenant_id: "TENANT001",
-        integration_key: "teams_connector",
-        integration_name: "Microsoft Teams連携",
-        integration_type: "WEBHOOK",
-        endpoint_url: "https://outlook.office.com/webhook/xxx/IncomingWebhook/yyy/zzz",
-        auth_type: "NONE",
-        auth_config: null,
-        connection_config: "{"title": "スキル管理システム", "theme_color": "0078D4"}",
-        request_headers: "{"Content-Type": "application/json"}",
-        timeout_seconds: 30,
-        retry_count: 3,
-        retry_interval: 5,
-        rate_limit_per_minute: 30,
-        is_enabled: true,
-        health_check_url: null,
-        last_health_check: null,
-        health_status: "UNKNOWN",
-        is_deleted: null,
-      },
-      {
-        id: "IC003",
-        tenant_id: "TENANT001",
-        integration_key: "smtp_server",
-        integration_name: "メール送信サーバー",
-        integration_type: "SMTP",
-        endpoint_url: "smtp.company.com:587",
-        auth_type: "BASIC",
-        auth_config: "{"username": "noreply@company.com", "password": "encrypted_password"}",
-        connection_config: "{"use_tls": true, "use_ssl": false, "from_address": "noreply@company.com", "from_name": "スキル管理システム"}",
-        request_headers: null,
-        timeout_seconds: 60,
-        retry_count: 2,
-        retry_interval: 10,
-        rate_limit_per_minute: 100,
-        is_enabled: true,
-        health_check_url: null,
-        last_health_check: "2025-06-01 19:00:00",
-        health_status: "HEALTHY",
-        is_deleted: null,
-      },
-    ],
-  })
-  
-  // SYS_MasterData_sample_data.sql
-  console.log('📊 SYS_MasterDataデータを投入中...')
-  await prisma.masterData.createMany({
-    data: [
-      {
-        master_key: "SYSTEM.MAX_LOGIN_ATTEMPTS",
-        master_category: "SYSTEM",
-        master_name: "最大ログイン試行回数",
-        master_value: 5,
-        data_type: "INTEGER",
-        default_value: 3,
-        validation_rule: "^[1-9][0-9]*$",
-        is_system_managed: true,
-        is_editable: true,
-        display_order: 1,
-        description: "ログイン失敗時の最大試行回数。この回数を超えるとアカウントがロックされます。",
-        effective_from: "2024-01-01",
-        effective_to: null,
-        last_modified_by: "system_admin",
-        last_modified_at: "2024-01-01 10:00:00",
-        version: 1,
-        id: null,
-        is_deleted: null,
-      },
-      {
-        master_key: "SYSTEM.SESSION_TIMEOUT_MINUTES",
-        master_category: "SYSTEM",
-        master_name: "セッションタイムアウト時間",
-        master_value: 30,
-        data_type: "INTEGER",
-        default_value: 30,
-        validation_rule: "^[1-9][0-9]*$",
-        is_system_managed: true,
-        is_editable: true,
-        display_order: 2,
-        description: "ユーザーセッションの有効時間（分）。この時間を過ぎると自動的にログアウトされます。",
-        effective_from: "2024-01-01",
-        effective_to: null,
-        last_modified_by: "system_admin",
-        last_modified_at: "2024-01-01 10:00:00",
-        version: 1,
-        id: null,
-        is_deleted: null,
-      },
-      {
-        master_key: "CODE.SKILL_LEVELS",
-        master_category: "CODE",
-        master_name: "スキルレベル定義",
-        master_value: "{"1":"初級","2":"中級","3":"上級","4":"エキスパート","5":"マスター"}",
-        data_type: "JSON",
-        default_value: "{"1":"初級","2":"中級","3":"上級","4":"エキスパート","5":"マスター"}",
-        validation_rule: null,
-        is_system_managed: false,
-        is_editable: true,
-        display_order: 1,
-        description: "スキル評価で使用するレベル定義。1-5の数値とその意味を定義します。",
-        effective_from: "2024-01-01",
-        effective_to: null,
-        last_modified_by: "admin_user",
-        last_modified_at: "2024-01-01 10:00:00",
-        version: 1,
-        id: null,
-        is_deleted: null,
-      },
-      {
-        master_key: "CONFIG.BACKUP_RETENTION_DAYS",
-        master_category: "CONFIG",
-        master_name: "バックアップ保持日数",
-        master_value: 90,
-        data_type: "INTEGER",
-        default_value: 30,
-        validation_rule: "^[1-9][0-9]*$",
-        is_system_managed: true,
-        is_editable: true,
-        display_order: 1,
-        description: "フルバックアップファイルの保持期間（日数）。この期間を過ぎたバックアップは自動削除されます。",
-        effective_from: "2024-01-01",
-        effective_to: null,
-        last_modified_by: "system_admin",
-        last_modified_at: "2024-01-01 10:00:00",
-        version: 1,
-        id: null,
-        is_deleted: null,
-      },
-    ],
-  })
-  
-  // SYS_SkillIndex_sample_data.sql
-  console.log('📊 SYS_SkillIndexデータを投入中...')
-  await prisma.skillIndex.createMany({
-    data: [
-      {
-        id: "SI001",
-        tenant_id: "TENANT001",
-        skill_id: "SKILL001",
-        index_type: "FULLTEXT",
-        search_term: "Java",
-        normalized_term: "java",
-        relevance_score: 1.0,
-        frequency_weight: 1.0,
-        position_weight: 1.0,
-        language_code: "ja",
-        source_field: "NAME",
-        is_active: true,
-        search_count: 150,
-        last_searched_at: "2025-06-01 18:30:00",
-        index_updated_at: "2025-06-01 10:00:00",
-        is_deleted: null,
-      },
-      {
-        id: "SI002",
-        tenant_id: "TENANT001",
-        skill_id: "SKILL001",
-        index_type: "KEYWORD",
-        search_term: "プログラミング",
-        normalized_term: "プログラミング",
-        relevance_score: 0.8,
-        frequency_weight: 0.75,
-        position_weight: 0.9,
-        language_code: "ja",
-        source_field: "DESCRIPTION",
-        is_active: true,
-        search_count: 85,
-        last_searched_at: "2025-06-01 17:45:00",
-        index_updated_at: "2025-06-01 10:00:00",
-        is_deleted: null,
-      },
-      {
-        id: "SI003",
-        tenant_id: "TENANT001",
-        skill_id: "SKILL001",
-        index_type: "SYNONYM",
-        search_term: "ジャバ",
-        normalized_term: "ジャバ",
-        relevance_score: 0.9,
-        frequency_weight: 0.5,
-        position_weight: 1.0,
-        language_code: "ja",
-        source_field: "KEYWORD",
-        is_active: true,
-        search_count: 25,
-        last_searched_at: "2025-06-01 16:20:00",
-        index_updated_at: "2025-06-01 10:00:00",
-        is_deleted: null,
-      },
-    ],
-  })
-  
-  // SYS_SkillMatrix_sample_data.sql
-  console.log('📊 SYS_SkillMatrixデータを投入中...')
-  await prisma.skillMatrix.createMany({
-    data: [
-      {
-        employee_id: "EMP001",
-        skill_id: "SKILL001",
-        skill_level: 3,
-        self_assessment: 3,
-        manager_assessment: 3,
-        peer_assessment: 2,
-        assessment_date: "2024-01-15",
-        evidence_url: "https://example.com/project/web-app",
-        notes: "Webアプリケーション開発プロジェクトでReactを使用",
-        next_target_level: 4,
-        target_date: "2024-06-30",
-        id: null,
-        is_deleted: null,
-      },
-      {
-        employee_id: "EMP001",
-        skill_id: "SKILL002",
-        skill_level: 2,
-        self_assessment: 2,
-        manager_assessment: 2,
-        peer_assessment: 3,
-        assessment_date: "2024-01-15",
-        evidence_url: null,
-        notes: "基本的なPython開発は可能、フレームワーク経験が少ない",
-        next_target_level: 3,
-        target_date: "2024-09-30",
-        id: null,
-        is_deleted: null,
-      },
-    ],
-  })
-  
-  // SYS_SystemLog_sample_data.sql
-  console.log('📊 SYS_SystemLogデータを投入中...')
-  await prisma.systemLog.createMany({
-    data: [
-      {
-        log_level: "INFO",
-        log_category: "AUTH",
-        message: "ユーザーログイン成功",
-        user_id: "user001",
-        session_id: "sess_abc123",
-        ip_address: "192.168.1.100",
-        user_agent: "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36",
-        request_url: "/api/auth/login",
-        request_method: "POST",
-        response_status: 200,
-        response_time: 150,
-        error_code: null,
-        stack_trace: null,
-        request_body: null,
-        response_body: null,
-        correlation_id: "corr_xyz789",
-        component_name: "AuthService",
-        thread_name: "http-nio-8080-exec-1",
-        server_name: "app-server-01",
-        id: null,
-        is_deleted: null,
-      },
-      {
-        log_level: "ERROR",
-        log_category: "API",
-        message: "データベース接続エラー",
-        user_id: "user002",
-        session_id: "sess_def456",
-        ip_address: "192.168.1.101",
-        user_agent: "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36",
-        request_url: "/api/employees",
-        request_method: "GET",
-        response_status: 500,
-        response_time: 5000,
-        error_code: "DB_CONNECTION_ERROR",
-        stack_trace: "java.sql.SQLException: Connection timeout...",
-        request_body: null,
-        response_body: null,
-        correlation_id: "corr_abc456",
-        component_name: "EmployeeService",
-        thread_name: "http-nio-8080-exec-2",
-        server_name: "app-server-02",
-        id: null,
-        is_deleted: null,
-      },
-    ],
-  })
-  
-  // SYS_TenantUsage_sample_data.sql
-  console.log('📊 SYS_TenantUsageデータを投入中...')
-  await prisma.tenantUsage.createMany({
-    data: [
-      {
-        usage_date: "2024-01-15",
-        tenant_id: "TENANT001",
-        active_users: 25,
-        total_logins: 47,
-        api_requests: 15420,
-        data_storage_mb: 1024.5,
-        file_storage_mb: 512.25,
-        backup_storage_mb: 256.75,
-        cpu_usage_minutes: 180.5,
-        memory_usage_mb_hours: 2048.0,
-        network_transfer_mb: 128.3,
-        report_generations: 12,
-        skill_assessments: 35,
-        notification_sent: 8,
-        peak_concurrent_users: 18,
-        peak_time: "14:30:00",
-        error_count: 2,
-        response_time_avg_ms: 245.5,
-        uptime_percentage: 99.95,
-        billing_amount: 1250.0,
-        collection_timestamp: "2024-01-16 02:00:00",
-        id: null,
-        is_deleted: null,
-      },
-      {
-        usage_date: "2024-01-15",
-        tenant_id: "TENANT002",
-        active_users: 8,
-        total_logins: 12,
-        api_requests: 3240,
-        data_storage_mb: 256.25,
-        file_storage_mb: 128.5,
-        backup_storage_mb: 64.25,
-        cpu_usage_minutes: 45.25,
-        memory_usage_mb_hours: 512.0,
-        network_transfer_mb: 32.1,
-        report_generations: 3,
-        skill_assessments: 8,
-        notification_sent: 2,
-        peak_concurrent_users: 6,
-        peak_time: "10:15:00",
-        error_count: 0,
-        response_time_avg_ms: 198.25,
-        uptime_percentage: 100.0,
-        billing_amount: 320.0,
-        collection_timestamp: "2024-01-16 02:00:00",
-        id: null,
-        is_deleted: null,
-      },
-    ],
-  })
-  
-  // SYS_TokenStore_sample_data.sql
-  console.log('📊 SYS_TokenStoreデータを投入中...')
-  await prisma.tokenStore.createMany({
-    data: [
-      {
-        id: "TS001",
-        tenant_id: "TENANT001",
-        user_id: "USER001",
-        token_type: "ACCESS",
-        token_value: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
-        token_hash: "a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6q7r8s9t0",
-        expires_at: "2025-06-01 20:00:00",
-        issued_at: "2025-06-01 19:00:00",
-        last_used_at: "2025-06-01 19:30:00",
-        client_ip: "192.168.1.100",
-        user_agent: "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36",
-        device_fingerprint: "fp_abc123def456",
-        scope: "["read:profile", "write:skills", "read:goals"]",
-        is_revoked: false,
-        revoked_at: null,
-        revoked_reason: null,
-        is_deleted: null,
-      },
-      {
-        id: "TS002",
-        tenant_id: "TENANT001",
-        user_id: "USER001",
-        token_type: "REFRESH",
-        token_value: "rt_xyz789abc123def456ghi789jkl012mno345",
-        token_hash: "b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6q7r8s9t0u1",
-        expires_at: "2025-06-08 19:00:00",
-        issued_at: "2025-06-01 19:00:00",
-        last_used_at: null,
-        client_ip: "192.168.1.100",
-        user_agent: "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36",
-        device_fingerprint: "fp_abc123def456",
-        scope: "["refresh"]",
-        is_revoked: false,
-        revoked_at: null,
-        revoked_reason: null,
-        is_deleted: null,
-      },
-    ],
-  })
-  
-  // TRN_EmployeeSkillGrade_sample_data.sql
-  console.log('📊 TRN_EmployeeSkillGradeデータを投入中...')
-  await prisma.employeeSkillGrade.createMany({
-    data: [
-      {
-        employee_id: "EMP000001",
-        job_type_id: "JOB001",
-        skill_grade: "A",
-        skill_level: 4,
-        effective_date: "2024-04-01",
-        expiry_date: null,
-        evaluation_date: "2024-03-15",
-        evaluator_id: "EMP000010",
-        evaluation_comment: "優秀な技術力と指導力を発揮している",
-        certification_flag: true,
-        next_evaluation_date: "2025-04-01",
-        id: null,
-        is_deleted: null,
-        tenant_id: null,
-        created_at: null,
-        updated_at: null,
-        created_by: null,
-        updated_by: null,
-      },
-      {
-        employee_id: "EMP000002",
-        job_type_id: "JOB002",
-        skill_grade: "B",
-        skill_level: 3,
-        effective_date: "2024-04-01",
-        expiry_date: null,
-        evaluation_date: "2024-03-20",
-        evaluator_id: "EMP000001",
-        evaluation_comment: "着実にスキルアップしており、次のレベルが期待される",
-        certification_flag: true,
-        next_evaluation_date: "2025-04-01",
-        id: null,
-        is_deleted: null,
-        tenant_id: null,
-        created_at: null,
-        updated_at: null,
-        created_by: null,
-        updated_by: null,
-      },
-      {
-        employee_id: "EMP000001",
-        job_type_id: "JOB001",
-        skill_grade: "B",
-        skill_level: 3,
-        effective_date: "2023-04-01",
-        expiry_date: "2024-03-31",
-        evaluation_date: "2023-03-15",
-        evaluator_id: "EMP000010",
-        evaluation_comment: "前年度からの成長が顕著",
-        certification_flag: true,
-        next_evaluation_date: "2024-04-01",
-        id: null,
-        is_deleted: null,
-        tenant_id: null,
-        created_at: null,
-        updated_at: null,
-        created_by: null,
-        updated_by: null,
-      },
-    ],
-  })
-  
-  // TRN_GoalProgress_sample_data.sql
-  console.log('📊 TRN_GoalProgressデータを投入中...')
-  await prisma.goalProgress.createMany({
-    data: [
-      {
-        goal_id: "GOAL000001",
-        employee_id: "EMP000001",
-        goal_title: "Java技術習得",
-        goal_description: "Spring Frameworkを使用したWebアプリケーション開発技術の習得",
-        goal_category: "SKILL",
-        goal_type: "QUALITATIVE",
-        priority_level: "HIGH",
-        target_value: null,
-        current_value: null,
-        unit: null,
-        start_date: "2025-01-01",
-        target_date: "2025-12-31",
-        progress_rate: 50.0,
-        achievement_status: "IN_PROGRESS",
-        supervisor_id: "EMP000010",
-        approval_status: "APPROVED",
-        approved_at: "2025-01-05 10:00:00",
-        approved_by: "EMP000010",
-        completion_date: null,
-        achievement_rate: null,
-        self_evaluation: null,
-        supervisor_evaluation: null,
-        evaluation_comments: null,
-        related_career_plan_id: "CP000001",
-        related_skill_items: "["JAVA", "SPRING", "WEB_DEVELOPMENT"]",
-        milestones: "["基礎学習完了", "実践プロジェクト参加", "技術認定取得"]",
-        obstacles: null,
-        support_needed: "外部研修参加、メンター指導",
-        last_updated_at: "2025-06-01 09:00:00",
-        next_review_date: "2025-07-01",
-        id: null,
-        is_deleted: null,
-        tenant_id: null,
-        created_at: null,
-        updated_at: null,
-        created_by: null,
-        updated_by: null,
-      },
-      {
-        goal_id: "GOAL000002",
-        employee_id: "EMP000002",
-        goal_title: "売上目標達成",
-        goal_description: "第2四半期の個人売上目標1000万円の達成",
-        goal_category: "BUSINESS",
-        goal_type: "QUANTITATIVE",
-        priority_level: "HIGH",
-        target_value: 10000000.0,
-        current_value: 6500000.0,
-        unit: "円",
-        start_date: "2025-04-01",
-        target_date: "2025-06-30",
-        progress_rate: 65.0,
-        achievement_status: "IN_PROGRESS",
-        supervisor_id: "EMP000011",
-        approval_status: "APPROVED",
-        approved_at: "2025-03-25 14:00:00",
-        approved_by: "EMP000011",
-        completion_date: null,
-        achievement_rate: null,
-        self_evaluation: null,
-        supervisor_evaluation: null,
-        evaluation_comments: null,
-        related_career_plan_id: null,
-        related_skill_items: "["SALES", "NEGOTIATION", "CUSTOMER_MANAGEMENT"]",
-        milestones: "["4月目標達成", "5月目標達成", "6月目標達成"]",
-        obstacles: "["競合他社の価格競争", "新規顧客開拓の困難"]",
-        support_needed: "マーケティング支援、価格戦略見直し",
-        last_updated_at: "2025-06-01 17:00:00",
-        next_review_date: "2025-06-15",
-        id: null,
-        is_deleted: null,
-        tenant_id: null,
-        created_at: null,
-        updated_at: null,
-        created_by: null,
-        updated_by: null,
-      },
-      {
-        goal_id: "GOAL000003",
-        employee_id: "EMP000003",
-        goal_title: "チームリーダー昇進",
-        goal_description: "リーダーシップスキル向上とチーム管理経験の積み重ね",
-        goal_category: "CAREER",
-        goal_type: "MILESTONE",
-        priority_level: "MEDIUM",
-        target_value: null,
-        current_value: null,
-        unit: null,
-        start_date: "2025-01-01",
-        target_date: "2025-12-31",
-        progress_rate: 30.0,
-        achievement_status: "IN_PROGRESS",
-        supervisor_id: "EMP000012",
-        approval_status: "APPROVED",
-        approved_at: "2025-01-10 11:00:00",
-        approved_by: "EMP000012",
-        completion_date: null,
-        achievement_rate: null,
-        self_evaluation: null,
-        supervisor_evaluation: null,
-        evaluation_comments: null,
-        related_career_plan_id: "CP000002",
-        related_skill_items: "["LEADERSHIP", "TEAM_MANAGEMENT", "COMMUNICATION"]",
-        milestones: "["リーダーシップ研修受講", "プロジェクトリーダー経験", "昇進面談"]",
-        obstacles: null,
-        support_needed: "リーダーシップ研修、メンタリング",
-        last_updated_at: "2025-06-01 12:00:00",
-        next_review_date: "2025-08-01",
-        id: null,
-        is_deleted: null,
-        tenant_id: null,
-        created_at: null,
-        updated_at: null,
-        created_by: null,
-        updated_by: null,
-      },
-    ],
-  })
-  
-  // TRN_Notification_sample_data.sql
-  console.log('📊 TRN_Notificationデータを投入中...')
-  await prisma.notification.createMany({
-    data: [
-      {
-        notification_id: "NOTIF_001",
-        recipient_id: "EMP000001",
-        sender_id: null,
-        notification_type: "REMINDER",
-        notification_category: "CERTIFICATION",
-        priority_level: "HIGH",
-        title: "AWS認定資格の更新期限が近づいています",
-        message: "お持ちのAWS認定ソリューションアーキテクト資格の有効期限が30日後に迫っています。更新手続きをお忘れなく。",
-        message_format: "PLAIN",
-        action_url: "/certifications/renewal/CERT_AWS_001",
-        action_label: "更新手続きへ",
-        delivery_method: "EMAIL",
-        delivery_status: "DELIVERED",
-        sent_at: "2024-05-01 09:00:00",
-        delivered_at: "2024-05-01 09:01:23",
-        read_status: "READ",
-        read_at: "2024-05-01 10:30:45",
-        archived_at: null,
-        expiry_date: "2024-06-01",
-        retry_count: 0,
-        max_retry_count: 3,
-        last_retry_at: null,
-        error_message: null,
-        external_message_id: "email_12345",
-        template_id: "TMPL_CERT_RENEWAL",
-        template_variables: "{"certification_name": "AWS認定ソリューションアーキテクト", "days_until_expiry": 30}",
-        related_entity_type: "CERTIFICATION",
-        related_entity_id: "CERT_AWS_001",
-        batch_id: null,
-        user_agent: "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36",
-        ip_address: "192.168.1.100",
-        device_type: "PC",
-        is_bulk_notification: false,
-        personalization_data: "{"preferred_language": "ja", "timezone": "Asia/Tokyo"}",
-        id: null,
-        is_deleted: null,
-        tenant_id: null,
-        created_at: null,
-        updated_at: null,
-        created_by: null,
-        updated_by: null,
-      },
-      {
-        notification_id: "NOTIF_002",
-        recipient_id: "EMP000002",
-        sender_id: "EMP000010",
-        notification_type: "APPROVAL",
-        notification_category: "TRAINING",
-        priority_level: "NORMAL",
-        title: "研修参加申請が承認されました",
-        message: "申請いただいた「プロジェクトマネジメント基礎研修」への参加が承認されました。研修日程をご確認ください。",
-        message_format: "HTML",
-        action_url: "/training/details/TRN_PROG_001",
-        action_label: "研修詳細を確認",
-        delivery_method: "IN_APP",
-        delivery_status: "DELIVERED",
-        sent_at: "2024-04-15 14:30:00",
-        delivered_at: "2024-04-15 14:30:01",
-        read_status: "READ",
-        read_at: "2024-04-15 15:45:20",
-        archived_at: "2024-04-20 10:00:00",
-        expiry_date: null,
-        retry_count: 0,
-        max_retry_count: 3,
-        last_retry_at: null,
-        error_message: null,
-        external_message_id: null,
-        template_id: "TMPL_TRAINING_APPROVAL",
-        template_variables: "{"training_name": "プロジェクトマネジメント基礎研修", "approver_name": "田中部長"}",
-        related_entity_type: "TRAINING",
-        related_entity_id: "TRN_PROG_001",
-        batch_id: null,
-        user_agent: "Mozilla/5.0 (iPhone; CPU iPhone OS 14_0 like Mac OS X)",
-        ip_address: "192.168.1.101",
-        device_type: "MOBILE",
-        is_bulk_notification: false,
-        personalization_data: "{"preferred_language": "ja", "notification_sound": true}",
-        id: null,
-        is_deleted: null,
-        tenant_id: null,
-        created_at: null,
-        updated_at: null,
-        created_by: null,
-        updated_by: null,
-      },
-    ],
-  })
-  
-  // TRN_PDU_sample_data.sql
-  console.log('📊 TRN_PDUデータを投入中...')
-  await prisma.pDU.createMany({
-    data: [
-      {
-        pdu_id: "PDU_001",
-        employee_id: "EMP000001",
-        certification_id: "CERT_PMP_001",
-        activity_type: "TRAINING",
-        activity_name: "アジャイル開発手法研修",
-        activity_description: "スクラム・カンバンを中心としたアジャイル開発手法の実践研修",
-        provider_name: "アジャイル協会",
-        activity_date: "2024-03-15",
-        start_time: "09:00:00",
-        end_time: "17:00:00",
-        duration_hours: 8.0,
-        pdu_points: 8.0,
-        pdu_category: "TECHNICAL",
-        pdu_subcategory: "Development Methods",
-        location: "東京研修センター",
-        cost: 50000,
-        cost_covered_by: "COMPANY",
-        evidence_type: "CERTIFICATE",
-        evidence_file_path: "/evidence/pdu/PDU_001_certificate.pdf",
-        certificate_number: "AGILE-2024-001",
-        instructor_name: "山田講師",
-        learning_objectives: "アジャイル開発手法の理解と実践スキル習得",
-        learning_outcomes: "スクラム・カンバンの基礎理解、実際のプロジェクトへの適用方法を習得",
-        skills_developed: "["アジャイル開発", "スクラム", "カンバン", "チーム運営"]",
-        approval_status: "APPROVED",
-        approved_by: "EMP000020",
-        approval_date: "2024-03-20",
-        approval_comment: "PMP資格維持に適切なPDU活動として承認",
-        expiry_date: "2027-03-15",
-        is_recurring: false,
-        recurrence_pattern: null,
-        related_training_id: "TRN_HIS_003",
-        related_project_id: null,
-        id: null,
-        is_deleted: null,
-        tenant_id: null,
-        created_at: null,
-        updated_at: null,
-        created_by: null,
-        updated_by: null,
-      },
-      {
-        pdu_id: "PDU_002",
-        employee_id: "EMP000002",
-        certification_id: "CERT_AWS_001",
-        activity_type: "CONFERENCE",
-        activity_name: "AWS re:Invent 2024",
-        activity_description: "AWSの最新技術動向とベストプラクティスに関するカンファレンス",
-        provider_name: "Amazon Web Services",
-        activity_date: "2024-11-28",
-        start_time: null,
-        end_time: null,
-        duration_hours: 32.0,
-        pdu_points: 32.0,
-        pdu_category: "TECHNICAL",
-        pdu_subcategory: "Cloud Technologies",
-        location: "ラスベガス（オンライン参加）",
-        cost: 200000,
-        cost_covered_by: "COMPANY",
-        evidence_type: "ATTENDANCE",
-        evidence_file_path: "/evidence/pdu/PDU_002_attendance.pdf",
-        certificate_number: null,
-        instructor_name: null,
-        learning_objectives: "AWS最新技術の習得とクラウドアーキテクチャスキル向上",
-        learning_outcomes: "最新のAWSサービス理解、セキュリティベストプラクティス習得",
-        skills_developed: "["AWS最新技術", "クラウドセキュリティ", "サーバーレス", "機械学習"]",
-        approval_status: "APPROVED",
-        approved_by: "EMP000020",
-        approval_date: "2024-12-05",
-        approval_comment: "AWS認定維持に有効なPDU活動として承認",
-        expiry_date: "2027-11-28",
-        is_recurring: false,
-        recurrence_pattern: null,
-        related_training_id: null,
-        related_project_id: "PRJ_REC_002",
-        id: null,
-        is_deleted: null,
-        tenant_id: null,
-        created_at: null,
-        updated_at: null,
-        created_by: null,
-        updated_by: null,
-      },
-    ],
-  })
-  
-  // TRN_ProjectRecord_sample_data.sql
-  console.log('📊 TRN_ProjectRecordデータを投入中...')
-  await prisma.projectRecord.createMany({
-    data: [
-      {
-        project_record_id: "PRJ_REC_001",
-        employee_id: "EMP000001",
-        project_name: "顧客管理システム刷新プロジェクト",
-        project_code: "CRM2024_001",
-        client_name: "株式会社サンプル",
-        project_type: "DEVELOPMENT",
-        project_scale: "LARGE",
-        start_date: "2024-01-15",
-        end_date: "2024-12-31",
-        participation_rate: 80.0,
-        role_title: "テックリード",
-        responsibilities: "システム設計、開発チームリード、技術選定",
-        technologies_used: "["Java", "Spring Boot", "PostgreSQL", "React", "Docker"]",
-        skills_applied: "["システム設計", "チームマネジメント", "技術選定"]",
-        achievements: "予定より2週間早期リリース、性能要件120%達成",
-        challenges_faced: "レガシーシステムとの連携、短納期対応",
-        lessons_learned: "マイクロサービス設計の重要性、チーム間コミュニケーション",
-        team_size: 8,
-        budget_range: "UNDER_100M",
-        project_status: "COMPLETED",
-        evaluation_score: 4.5,
-        evaluation_comment: "技術リーダーシップを発揮し、プロジェクトを成功に導いた",
-        is_confidential: false,
-        is_public_reference: true,
-        id: null,
-        is_deleted: null,
-        tenant_id: null,
-        created_at: null,
-        updated_at: null,
-        created_by: null,
-        updated_by: null,
-      },
-      {
-        project_record_id: "PRJ_REC_002",
-        employee_id: "EMP000002",
-        project_name: "AI画像解析システム開発",
-        project_code: "AI2024_002",
-        client_name: "機密プロジェクト",
-        project_type: "RESEARCH",
-        project_scale: "MEDIUM",
-        start_date: "2024-03-01",
-        end_date: null,
-        participation_rate: 100.0,
-        role_title: "AIエンジニア",
-        responsibilities: "機械学習モデル開発、データ前処理、精度改善",
-        technologies_used: "["Python", "TensorFlow", "OpenCV", "AWS SageMaker"]",
-        skills_applied: "["機械学習", "画像処理", "データ分析"]",
-        achievements: "認識精度95%達成、処理速度30%向上",
-        challenges_faced: "学習データ不足、モデル精度向上",
-        lessons_learned: "データ品質の重要性、MLOpsの必要性",
-        team_size: 4,
-        budget_range: "UNDER_10M",
-        project_status: "ONGOING",
-        evaluation_score: null,
-        evaluation_comment: null,
-        is_confidential: true,
-        is_public_reference: false,
-        id: null,
-        is_deleted: null,
-        tenant_id: null,
-        created_at: null,
-        updated_at: null,
-        created_by: null,
-        updated_by: null,
-      },
-    ],
-  })
-  
-  // TRN_SkillEvidence_sample_data.sql
-  console.log('📊 TRN_SkillEvidenceデータを投入中...')
-  await prisma.skillEvidence.createMany({
-    data: [
-      {
-        evidence_id: "EVD_001",
-        employee_id: "EMP000001",
-        skill_id: "SKILL_JAVA_001",
-        evidence_type: "PROJECT",
-        evidence_title: "ECサイト基盤システム開発",
-        evidence_description: "大規模ECサイトのバックエンドシステムをJavaで設計・開発",
-        skill_level_demonstrated: "ADVANCED",
-        evidence_date: "2024-03-31",
-        validity_start_date: "2024-03-31",
-        validity_end_date: null,
-        file_path: "/evidence/EVD_001_project_summary.pdf",
-        file_type: "PDF",
-        file_size_kb: 2048,
-        external_url: "https://github.com/company/ecommerce-backend",
-        issuer_name: "株式会社サンプル",
-        issuer_type: "COMPANY",
-        certificate_number: null,
-        verification_method: "MANAGER",
-        verification_status: "VERIFIED",
-        verified_by: "EMP000010",
-        verification_date: "2024-04-05",
-        verification_comment: "高品質なコードと優れた設計により、システムの安定性と拡張性を実現",
-        related_project_id: "PRJ_REC_001",
-        related_training_id: null,
-        related_certification_id: null,
-        impact_score: 4.5,
-        complexity_level: "HIGH",
-        team_size: 8,
-        role_in_activity: "テックリード",
-        technologies_used: "["Java", "Spring Boot", "PostgreSQL", "Redis", "Docker"]",
-        achievements: "予定より2週間早期リリース、性能要件120%達成、バグ発生率0.1%以下",
-        lessons_learned: "大規模システムでのマイクロサービス設計、チーム間連携の重要性",
-        is_public: false,
-        is_portfolio_item: true,
-        tags: "["Java", "Spring Boot", "システム設計", "チームリード"]",
-        id: null,
-        is_deleted: null,
-        tenant_id: null,
-        created_at: null,
-        updated_at: null,
-        created_by: null,
-        updated_by: null,
-      },
-      {
-        evidence_id: "EVD_002",
-        employee_id: "EMP000002",
-        skill_id: "SKILL_AWS_001",
-        evidence_type: "CERTIFICATION",
-        evidence_title: "AWS認定ソリューションアーキテクト - アソシエイト",
-        evidence_description: "AWSクラウドサービスの設計・構築に関する認定資格",
-        skill_level_demonstrated: "INTERMEDIATE",
-        evidence_date: "2024-02-15",
-        validity_start_date: "2024-02-15",
-        validity_end_date: "2027-02-15",
-        file_path: "/evidence/EVD_002_aws_certificate.pdf",
-        file_type: "PDF",
-        file_size_kb: 512,
-        external_url: "https://aws.amazon.com/verification",
-        issuer_name: "Amazon Web Services",
-        issuer_type: "CERTIFICATION_BODY",
-        certificate_number: "AWS-SAA-2024-002",
-        verification_method: "AUTOMATIC",
-        verification_status: "VERIFIED",
-        verified_by: null,
-        verification_date: "2024-02-15",
-        verification_comment: "AWS公式認定により自動検証",
-        related_project_id: null,
-        related_training_id: "TRN_HIS_001",
-        related_certification_id: "CERT_AWS_001",
-        impact_score: 4.0,
-        complexity_level: "MEDIUM",
-        team_size: null,
-        role_in_activity: "受験者",
-        technologies_used: "["AWS", "EC2", "S3", "RDS", "Lambda"]",
-        achievements: "一発合格、スコア850点（合格ライン720点）",
-        lessons_learned: "クラウドアーキテクチャの設計原則、AWSサービスの適切な選択方法",
-        is_public: true,
-        is_portfolio_item: true,
-        tags: "["AWS", "クラウド", "認定資格", "アーキテクチャ"]",
-        id: null,
-        is_deleted: null,
-        tenant_id: null,
-        created_at: null,
-        updated_at: null,
-        created_by: null,
-        updated_by: null,
-      },
-    ],
-  })
-  
-  // TRN_SkillRecord_sample_data.sql
-  console.log('📊 TRN_SkillRecordデータを投入中...')
-  await prisma.skillRecord.createMany({
-    data: [
-      {
-        employee_id: "EMP000001",
-        skill_item_id: "SKILL001",
-        skill_level: 4,
-        self_assessment: 4,
-        manager_assessment: 3,
-        evidence_description: "Javaを使用したWebアプリケーション開発プロジェクトを3件担当",
-        acquisition_date: "2020-06-01",
-        last_used_date: "2025-05-30",
-        expiry_date: null,
-        certification_id: "CERT001",
-        skill_category_id: "CAT001",
-        assessment_date: "2025-04-01",
-        assessor_id: "EMP000010",
-        skill_status: "ACTIVE",
-        learning_hours: 120,
-        project_experience_count: 3,
-        id: null,
-        is_deleted: null,
-        tenant_id: null,
-        created_at: null,
-        updated_at: null,
-        created_by: null,
-        updated_by: null,
-      },
-      {
-        employee_id: "EMP000001",
-        skill_item_id: "SKILL002",
-        skill_level: 3,
-        self_assessment: 3,
-        manager_assessment: 3,
-        evidence_description: "AWS環境でのインフラ構築・運用経験",
-        acquisition_date: "2021-03-15",
-        last_used_date: "2025-05-25",
-        expiry_date: "2026-03-15",
-        certification_id: "CERT002",
-        skill_category_id: "CAT002",
-        assessment_date: "2025-04-01",
-        assessor_id: "EMP000010",
-        skill_status: "ACTIVE",
-        learning_hours: 80,
-        project_experience_count: 2,
-        id: null,
-        is_deleted: null,
-        tenant_id: null,
-        created_at: null,
-        updated_at: null,
-        created_by: null,
-        updated_by: null,
-      },
-    ],
-  })
-  
-  // TRN_TrainingHistory_sample_data.sql
-  console.log('📊 TRN_TrainingHistoryデータを投入中...')
-  await prisma.trainingHistory.createMany({
-    data: [
-      {
-        training_history_id: "TRN_HIS_001",
-        employee_id: "EMP000001",
-        training_program_id: "TRN_PROG_001",
-        training_name: "AWS認定ソリューションアーキテクト研修",
-        training_type: "EXTERNAL",
-        training_category: "TECHNICAL",
-        provider_name: "AWS Training",
-        instructor_name: "田中講師",
-        start_date: "2024-03-01",
-        end_date: "2024-03-03",
-        duration_hours: 24.0,
-        location: "東京研修センター",
-        cost: 150000,
-        cost_covered_by: "COMPANY",
-        attendance_status: "COMPLETED",
-        completion_rate: 100.0,
-        test_score: 85.0,
-        grade: "合格",
-        certificate_obtained: true,
-        certificate_number: "AWS-SAA-2024-001",
-        pdu_earned: 24.0,
-        skills_acquired: "["AWS設計", "クラウドアーキテクチャ", "セキュリティ設計"]",
-        learning_objectives: "AWSでのソリューション設計スキル習得",
-        learning_outcomes: "クラウドアーキテクチャ設計の基礎を習得、実践的な設計手法を学習",
-        feedback: "実践的な内容で非常に有用だった。講師の説明も分かりやすい。",
-        satisfaction_score: 4.5,
-        recommendation_score: 5.0,
-        follow_up_required: true,
-        follow_up_date: "2024-06-01",
-        manager_approval: true,
-        approved_by: "EMP000010",
-        id: null,
-        is_deleted: null,
-        tenant_id: null,
-        created_at: null,
-        updated_at: null,
-        created_by: null,
-        updated_by: null,
-      },
-      {
-        training_history_id: "TRN_HIS_002",
-        employee_id: "EMP000002",
-        training_program_id: null,
-        training_name: "プロジェクトマネジメント基礎",
-        training_type: "ONLINE",
-        training_category: "MANAGEMENT",
-        provider_name: "社内研修センター",
-        instructor_name: "佐藤部長",
-        start_date: "2024-02-15",
-        end_date: "2024-02-15",
-        duration_hours: 8.0,
-        location: "オンライン",
-        cost: 0,
-        cost_covered_by: "COMPANY",
-        attendance_status: "COMPLETED",
-        completion_rate: 100.0,
-        test_score: 92.0,
-        grade: "A",
-        certificate_obtained: true,
-        certificate_number: "PM-BASIC-2024-002",
-        pdu_earned: 8.0,
-        skills_acquired: "["プロジェクト計画", "リスク管理", "チームマネジメント"]",
-        learning_objectives: "プロジェクトマネジメントの基礎知識習得",
-        learning_outcomes: "PMBOKの基礎理解、実際のプロジェクト運営に活用可能な知識を習得",
-        feedback: "基礎から体系的に学べて良かった。実例が豊富で理解しやすい。",
-        satisfaction_score: 4.0,
-        recommendation_score: 4.0,
-        follow_up_required: false,
-        follow_up_date: null,
-        manager_approval: true,
-        approved_by: "EMP000010",
-        id: null,
-        is_deleted: null,
-        tenant_id: null,
-        created_at: null,
-        updated_at: null,
-        created_by: null,
-        updated_by: null,
-      },
-    ],
-  })
-  
-  // WRK_BatchJobLog_sample_data.sql
-  console.log('📊 WRK_BatchJobLogデータを投入中...')
-  await prisma.batchJobLog.createMany({
-    data: [
-      {
-        job_id: "JOB-001-20250601-001",
-        job_name: "スキル情報一括登録（2025年6月分）",
-        job_type: "SKILL_IMPORT",
-        status: "COMPLETED",
-        start_time: "2025-06-01 09:00:00",
-        end_time: "2025-06-01 09:15:30",
-        total_records: 500,
-        processed_records: 500,
-        success_records: 485,
-        error_records: 15,
-        error_details: "{"errors": [{"row": 23, "column": "skill_level", "message": "無効なスキルレベル値"}, {"row": 45, "column": "employee_id", "message": "存在しない社員ID"}]}",
-        input_file_path: "/uploads/skill_import_20250601.csv",
-        output_file_path: "/reports/skill_import_error_20250601.csv",
-        executed_by: "admin001",
-        progress_percentage: 100.0,
-        execution_environment: "batch-server-01:12345",
-        job_parameters: "{"delimiter": ",", "encoding": "UTF-8", "skip_header": true}",
-        id: null,
-        is_deleted: null,
-      },
-      {
-        job_id: "JOB-002-20250601-002",
-        job_name: "社員部署情報一括更新",
-        job_type: "BULK_UPDATE",
-        status: "RUNNING",
-        start_time: "2025-06-01 14:30:00",
-        end_time: null,
-        total_records: 1000,
-        processed_records: 650,
-        success_records: 640,
-        error_records: 10,
-        error_details: "{"errors": [{"row": 15, "column": "department_id", "message": "存在しない部署ID"}]}",
-        input_file_path: "/uploads/employee_update_20250601.csv",
-        output_file_path: null,
-        executed_by: "hr_manager001",
-        progress_percentage: 65.0,
-        execution_environment: "batch-server-02:23456",
-        job_parameters: "{"update_mode": "partial", "validate_only": false}",
-        id: null,
-        is_deleted: null,
-      },
-      {
-        job_id: "JOB-003-20250601-003",
-        job_name: "全社員スキルデータエクスポート",
-        job_type: "DATA_EXPORT",
-        status: "FAILED",
-        start_time: "2025-06-01 16:00:00",
-        end_time: "2025-06-01 16:05:15",
-        total_records: 0,
-        processed_records: 0,
-        success_records: 0,
-        error_records: 0,
-        error_details: "{"system_error": "データベース接続エラー", "error_code": "DB_CONNECTION_TIMEOUT"}",
-        input_file_path: null,
-        output_file_path: "/exports/skill_data_20250601.csv",
-        executed_by: "system_admin001",
-        progress_percentage: 0.0,
-        execution_environment: "batch-server-01:34567",
-        job_parameters: "{"format": "CSV", "include_deleted": false}",
-        id: null,
-        is_deleted: null,
-      },
-    ],
-  })
-  
-  console.log('✅ 初期データ投入が完了しました！')
+    }),
+  ])
+
+  console.log('✅ データベースの初期データ投入が完了しました！')
+  console.log('📋 投入されたデータ:')
+  console.log(`   - テナント: 1件`)
+  console.log(`   - 部署: ${departments.length}件`)
+  console.log(`   - 役職: ${positions.length}件`)
+  console.log(`   - 権限: ${roles.length}件`)
+  console.log(`   - 権限設定: ${permissions.length}件`)
+  console.log(`   - スキルカテゴリ: ${skillCategories.length}件`)
+  console.log(`   - スキル: ${skills.length}件`)
+  console.log(`   - 従業員: ${employees.length}件`)
+  console.log(`   - 認証情報: ${userAuths.length}件`)
+  console.log('')
+  console.log('🔐 ログイン情報:')
+  console.log('   管理者:')
+  console.log('     ユーザーID: admin@skill-report.local')
+  console.log('     パスワード: password')
+  console.log('   テストユーザー:')
+  console.log('     ユーザーID: test@skill-report.local')
+  console.log('     パスワード: password')
 }
 
-if (require.main === module) {
-  runSampleSeed().then(() => prisma.$disconnect())
-}
+main()
+  .then(async () => {
+    await prisma.$disconnect()
+  })
+  .catch(async (e) => {
+    console.error('❌ 初期データ投入中にエラーが発生しました:', e)
+    await prisma.$disconnect()
+    throw e
+  })
