+++ src/database/prisma/seed.ts.backup.20250606_155800
@@ -1,0 +1,41 @@
+// 要求仕様ID: PLT.1-DB.1 - データベース初期データ投入
+// 設計書: docs/design/database/data/ 配下のサンプルデータSQLファイル群
+// 自動生成日時: 2025-06-06 15:54:24
+import { PrismaClient } from '@prisma/client'
+
+const prisma = new PrismaClient()
+
+async function main() {
+  console.log('🌱 データベースの初期データ投入を開始します...')
+
+  try {
+
+    console.log('✅ データベースの初期データ投入が完了しました！')
+    console.log('📋 投入されたデータの詳細はログを確認してください')
+    console.log('')
+    console.log('🔐 ログイン情報:')
+    console.log('   管理者:')
+    console.log('     ユーザーID: admin@skill-report.local')
+    console.log('     パスワード: password')
+    console.log('   テストユーザー1:')
+    console.log('     ユーザーID: yamada.taro@company.com')
+    console.log('     パスワード: password')
+    console.log('   テストユーザー2:')
+    console.log('     ユーザーID: sato.hanako@company.com')
+    console.log('     パスワード: password')
+
+  } catch (error) {
+    console.error('❌ 初期データ投入中にエラーが発生しました:', error)
+    throw error
+  }
+}
+
+main()
+  .then(async () => {
+    await prisma.$disconnect()
+  })
+  .catch(async (e) => {
+    console.error('❌ 初期データ投入中にエラーが発生しました:', e)
+    await prisma.$disconnect()
+    throw e
+  })
